Dynamic Time Warping for Quaternion Time Series

This function evaluates the Dynamic Time Warping (DTW) distance between two quaternion time series (QTS).

Usage

DTW(
  qts1,
  qts2,
  resample = TRUE,
  disable_normalization = FALSE,
  distance_only = FALSE,
  step_pattern = dtw::symmetric2
)

Arguments

qts1

An object of class qts.

qts2

An object of class qts.

resample

A boolean specifying whether the QTS should be uniformly resampled on their domain before computing distances. Defaults to TRUE.

disable_normalization

A boolean specifying whether quaternion normalization should be disabled. Defaults to FALSE which ensures that we always deal with unit quaternions.

distance_only

A boolean specifying whether to only compute distance (no backtrack, faster). Defaults to FALSE.

step_pattern

A dtw::stepPattern specifying the local constraints on the warping path. Defaults to dtw::symmetric2 which uses symmetric and normalizable warping paths with no local slope constraints. See dtw::stepPattern for more information.

Value

An object of class dtw::dtw storing the dynamic time warping results.

Details

If no evaluation grid is provided, the function assumes that the two input QTS are evaluated on the same grid.

Examples

DTW(vespa64$igp[[1]], vespa64$igp[[2]])
#> 0.0797319 0.0698791 0.0133386  0.994275 0.0745706 0.0676322 0.0131288  0.994833
#> 0.0745706 0.0676322 0.0131288  0.994833 0.0693108 0.0645704 0.0126874  0.995422
#> 0.0693108 0.0645704 0.0126874  0.995422 0.0639757 0.0609104 0.0118434   0.99602
#> 0.0639757 0.0609104 0.0118434   0.99602 0.0594915 0.0574216 0.0107045  0.996518
#> 0.0594915 0.0574216 0.0107045  0.996518  0.0557187  0.0540293 0.00932125    0.99694
#>  0.0557187  0.0540293 0.00932125    0.99694  0.0527473  0.0507937 0.00771565   0.997285
#>  0.0527473  0.0507937 0.00771565   0.997285  0.0505641    0.04761 0.00583348   0.997568
#>  0.0505641    0.04761 0.00583348   0.997568  0.0488309  0.0443048 0.00365833   0.997817
#>  0.0488309  0.0443048 0.00365833   0.997817  0.0472977  0.0407108 0.00125483    0.99805
#>  0.0472977  0.0407108 0.00125483    0.99805   0.0455235   0.0368852 -0.00124754    0.998281
#>   0.0455235   0.0368852 -0.00124754    0.998281   0.0433048   0.0330652 -0.00365939    0.998508
#>   0.0433048   0.0330652 -0.00365939    0.998508   0.0406827   0.0293761 -0.00595589    0.998722
#>   0.0406827   0.0293761 -0.00595589    0.998722   0.0375039   0.0258003 -0.00809226    0.998931
#>   0.0375039   0.0258003 -0.00809226    0.998931   0.0338919   0.0224858 -0.00985882    0.999124
#>   0.0338919   0.0224858 -0.00985882    0.999124  0.0295721  0.0191493 -0.0112784   0.999316
#>  0.0295721  0.0191493 -0.0112784   0.999316  0.0246661  0.0158799 -0.0123314   0.999494
#>  0.0246661  0.0158799 -0.0123314   0.999494  0.0195056  0.0127247 -0.0132062   0.999642
#>  0.0195056  0.0127247 -0.0132062   0.999642  0.0141261  0.0095846 -0.0140144   0.999756
#>  0.0141261  0.0095846 -0.0140144   0.999756 0.00851007 0.00654356 -0.0148724   0.999832
#> 0.00851007 0.00654356 -0.0148724   0.999832 0.00263847 0.00368908 -0.0158786   0.999864
#> 0.00263847 0.00368908 -0.0158786   0.999864 -0.00346152  0.00105427  -0.0170618    0.999848
#> -0.00346152  0.00105427  -0.0170618    0.999848 -0.00949419 -0.00122162  -0.0183589    0.999786
#> -0.00949419 -0.00122162  -0.0183589    0.999786  -0.0154556 -0.00317885  -0.0197861     0.99968
#>  -0.0154556 -0.00317885  -0.0197861     0.99968  -0.0211829 -0.00468279  -0.0212178    0.999539
#>  -0.0211829 -0.00468279  -0.0212178    0.999539  -0.0267815 -0.00579118  -0.0225796    0.999369
#>  -0.0267815 -0.00579118  -0.0225796    0.999369  -0.0321474 -0.00644995   -0.023764     0.99918
#>  -0.0321474 -0.00644995   -0.023764     0.99918  -0.0372149 -0.00684313  -0.0247801    0.998977
#>  -0.0372149 -0.00684313  -0.0247801    0.998977  -0.0423338 -0.00718711  -0.0257854    0.998745
#>  -0.0423338 -0.00718711  -0.0257854    0.998745   -0.047525 -0.00751707  -0.0267669    0.998483
#>   -0.047525 -0.00751707  -0.0267669    0.998483 -0.0526777 -0.0078594  -0.027707   0.998196
#> -0.0526777 -0.0078594  -0.027707   0.998196   -0.057765 -0.00817027  -0.0286047    0.997887
#>   -0.057765 -0.00817027  -0.0286047    0.997887  -0.0630269 -0.00837104  -0.0294466    0.997542
#>  -0.0630269 -0.00837104  -0.0294466    0.997542  -0.0683562 -0.00842349  -0.0302018    0.997168
#>  -0.0683562 -0.00842349  -0.0302018    0.997168  -0.0740528 -0.00840281   -0.030905     0.99674
#>  -0.0740528 -0.00840281   -0.030905     0.99674  -0.0797429 -0.00845896  -0.0315681     0.99628
#>  -0.0797429 -0.00845896  -0.0315681     0.99628  -0.0850624 -0.00867096  -0.0321786    0.995818
#>  -0.0850624 -0.00867096  -0.0321786    0.995818   -0.089533 -0.00910213   -0.032703    0.995405
#>   -0.089533 -0.00910213   -0.032703    0.995405  -0.0929019 -0.00978977  -0.0331882    0.995074
#>  -0.0929019 -0.00978977  -0.0331882    0.995074  -0.095039 -0.0107645 -0.0336473   0.994847
#>  -0.095039 -0.0107645 -0.0336473   0.994847 -0.0960733 -0.0118675 -0.0340958   0.994719
#> -0.0960733 -0.0118675 -0.0340958   0.994719  -0.096135 -0.0130157 -0.0345222   0.994684
#>  -0.096135 -0.0130157 -0.0345222   0.994684 -0.0952552 -0.0141094 -0.0349085   0.994741
#> -0.0952552 -0.0141094 -0.0349085   0.994741 -0.0938689 -0.0150725 -0.0351549    0.99485
#> -0.0938689 -0.0150725 -0.0351549    0.99485 -0.0923685 -0.0159798 -0.0352098   0.994974
#> -0.0923685 -0.0159798 -0.0352098   0.994974 -0.0911558 -0.0168873 -0.0349284   0.995081
#> -0.0911558 -0.0168873 -0.0349284   0.995081 -0.0899919 -0.0178851 -0.0342654   0.995192
#> -0.0899919 -0.0178851 -0.0342654   0.995192 -0.0886048 -0.0192397 -0.0333316   0.995323
#> -0.0886048 -0.0192397 -0.0333316   0.995323 -0.0870326 -0.0210119  -0.032204   0.995463
#> -0.0870326 -0.0210119  -0.032204   0.995463 -0.0854349 -0.0235867 -0.0307271    0.99559
#> -0.0854349 -0.0235867 -0.0307271    0.99559 -0.0839029 -0.0269733 -0.0289475   0.995688
#> -0.0839029 -0.0269733 -0.0289475   0.995688 -0.0825558 -0.0312877 -0.0266873   0.995738
#> -0.0825558 -0.0312877 -0.0266873   0.995738 -0.0815756 -0.0366626 -0.0235873   0.995713
#> -0.0815756 -0.0366626 -0.0235873   0.995713 -0.0806604 -0.0421512 -0.0194757    0.99566
#> -0.0806604 -0.0421512 -0.0194757    0.99566 -0.0795706 -0.0472723 -0.0145172   0.995602
#> -0.0795706 -0.0472723 -0.0145172   0.995602  -0.0781378  -0.0516994 -0.00905205     0.99556
#>  -0.0781378  -0.0516994 -0.00905205     0.99556  -0.0762388  -0.0553149 -0.00375525    0.995547
#>  -0.0762388  -0.0553149 -0.00375525    0.995547  -0.0740653  -0.0582896 0.000918781    0.995548
#>  -0.0740653  -0.0582896 0.000918781    0.995548 -0.0718354 -0.0604315 0.00475816   0.995573
#> -0.0718354 -0.0604315 0.00475816   0.995573 -0.0697407 -0.0616088 0.00755043   0.995632
#> -0.0697407 -0.0616088 0.00755043   0.995632 -0.0678403 -0.0617618 0.00979725   0.995734
#> -0.0678403 -0.0617618 0.00979725   0.995734 -0.0661331 -0.0608666  0.0116406   0.995885
#> -0.0661331 -0.0608666  0.0116406   0.995885 -0.0646695 -0.0588047  0.0131003   0.996086
#> -0.0646695 -0.0588047  0.0131003   0.996086 -0.0633517 -0.0555818  0.0146001   0.996335
#> -0.0633517 -0.0555818  0.0146001   0.996335 -0.0621399 -0.0510562  0.0162449   0.996628
#> -0.0621399 -0.0510562  0.0162449   0.996628 -0.0608728 -0.0451877  0.0180479   0.996959
#> -0.0608728 -0.0451877  0.0180479   0.996959 -0.0590487 -0.0385883    0.01997   0.997309
#> -0.0590487 -0.0385883    0.01997   0.997309 -0.0565467  -0.031285  0.0220677   0.997666
#> -0.0565467  -0.031285  0.0220677   0.997666 -0.0534494  -0.023539   0.024311   0.997997
#> -0.0534494  -0.023539   0.024311   0.997997  -0.049874 -0.0157996  0.0265733   0.998277
#>  -0.049874 -0.0157996  0.0265733   0.998277  -0.0456449 -0.00801856   0.0288406    0.998509
#>  -0.0456449 -0.00801856   0.0288406    0.998509   -0.0408634 -0.000356516    0.0310013     0.998684
#>   -0.0408634 -0.000356516    0.0310013     0.998684 -0.0358065 0.00716518  0.0329859   0.998789
#> -0.0358065 0.00716518  0.0329859   0.998789 -0.0304935  0.0146038   0.034792   0.998823
#> -0.0304935  0.0146038   0.034792   0.998823 -0.0249401  0.0215898  0.0362626   0.998798
#> -0.0249401  0.0215898  0.0362626   0.998798 -0.0189391  0.0281274  0.0373484   0.998727
#> -0.0189391  0.0281274  0.0373484   0.998727 -0.0127508  0.0340646  0.0380724   0.998613
#> -0.0127508  0.0340646  0.0380724   0.998613 -0.00632422   0.0393325   0.0384625    0.998466
#> -0.00632422   0.0393325   0.0384625    0.998466 -0.000108763    0.0435897    0.0384339      0.99831
#> -0.000108763    0.0435897    0.0384339      0.99831 0.00617018  0.0471212  0.0380387   0.998146
#> 0.00617018  0.0471212  0.0380387   0.998146 0.0124052 0.0499296 0.0372287  0.997982
#> 0.0124052 0.0499296 0.0372287  0.997982 0.0183735 0.0518609 0.0360666  0.997834
#> 0.0183735 0.0518609 0.0360666  0.997834 0.0239493 0.0530305  0.034622  0.997705
#> 0.0239493 0.0530305  0.034622  0.997705 0.0292805 0.0534231 0.0329226  0.997599
#> 0.0292805 0.0534231 0.0329226  0.997599 0.0346959 0.0531194 0.0309587  0.997505
#> 0.0346959 0.0531194 0.0309587  0.997505 0.0401971 0.0523779  0.029039  0.997395
#> 0.0401971 0.0523779  0.029039  0.997395 0.0456276 0.0513486 0.0271414  0.997269
#> 0.0456276 0.0513486 0.0271414  0.997269 0.0506394 0.0502362 0.0254119  0.997129
#> 0.0506394 0.0502362 0.0254119  0.997129 0.0552249 0.0492102 0.0237877  0.996977
#> 0.0552249 0.0492102 0.0237877  0.996977 0.0592191 0.0485406  0.022206  0.996817
#> 0.0592191 0.0485406  0.022206  0.996817 0.0624644 0.0483189 0.0206319  0.996663
#> 0.0624644 0.0483189 0.0206319  0.996663 0.0649023 0.0487759 0.0191541  0.996515
#> 0.0649023 0.0487759 0.0191541  0.996515 0.0667587 0.0498722 0.0178024  0.996363
#> 0.0667587 0.0498722 0.0178024  0.996363 0.0681882  0.051669 0.0167233  0.996193
#> 0.0681882  0.051669 0.0167233  0.996193 0.0695096 0.0539529 0.0159898  0.995993
#> 0.0695096 0.0539529 0.0159898  0.995993 0.0713586  0.056374 0.0155501  0.995735
#> 0.0713586  0.056374 0.0155501  0.995735 0.0736909 0.0586148 0.0151908  0.995441
#> 0.0736909 0.0586148 0.0151908  0.995441  0.075707 0.0609334 0.0147754  0.995157
#>  0.075707 0.0609334 0.0147754  0.995157 0.0775935  0.063441  0.014245  0.994863
#> 0.0775935  0.063441  0.014245  0.994863 0.0791957 0.0664329 0.0136903  0.994549
#> 0.0791957 0.0664329 0.0136903  0.994549 0.0797319 0.0698791 0.0133386  0.994275
#>  0.0828199  0.0736952 0.00993852   0.993786 0.0772691 0.0711958 0.0103556  0.994411
#> 0.0772691 0.0711958 0.0103556  0.994411 0.0717107 0.0682931 0.0105934  0.995028
#> 0.0717107 0.0682931 0.0105934  0.995028 0.0659838 0.0648941 0.0104718  0.995653
#> 0.0659838 0.0648941 0.0104718  0.995653  0.060324  0.061241 0.0100535  0.996248
#>  0.060324  0.061241 0.0100535  0.996248  0.0552466  0.0575625 0.00926533   0.996769
#>  0.0552466  0.0575625 0.00926533   0.996769  0.0514188  0.0541744 0.00810266   0.997174
#>  0.0514188  0.0541744 0.00810266   0.997174  0.0484918  0.0508021 0.00654867   0.997509
#>  0.0484918  0.0508021 0.00654867   0.997509  0.0462156  0.0472326 0.00460901   0.997804
#>  0.0462156  0.0472326 0.00460901   0.997804  0.0444218  0.0434468 0.00236554   0.998065
#>  0.0444218  0.0434468 0.00236554   0.998065    0.0430048    0.0396484 -2.14885e-05     0.998288
#>    0.0430048    0.0396484 -2.14885e-05     0.998288   0.0416644   0.0359638 -0.00239914    0.998481
#>   0.0416644   0.0359638 -0.00239914    0.998481   0.0400693   0.0324726 -0.00468125    0.998658
#>   0.0400693   0.0324726 -0.00468125    0.998658   0.0378909   0.0291498 -0.00678326    0.998834
#>   0.0378909   0.0291498 -0.00678326    0.998834   0.0349994   0.0259801 -0.00864534    0.999012
#>   0.0349994   0.0259801 -0.00864534    0.999012  0.0314968  0.0229827 -0.0102005   0.999188
#>  0.0314968  0.0229827 -0.0102005   0.999188  0.0275948  0.0201301 -0.0114319   0.999351
#>  0.0275948  0.0201301 -0.0114319   0.999351  0.0230019  0.0172827 -0.0123409    0.99951
#>  0.0230019  0.0172827 -0.0123409    0.99951  0.0178166  0.0143603 -0.0131072   0.999652
#>  0.0178166  0.0143603 -0.0131072   0.999652  0.0121847  0.0114108 -0.0138099   0.999765
#>  0.0121847  0.0114108 -0.0138099   0.999765 0.00619608 0.00840595 -0.0145856   0.999839
#> 0.00619608 0.00840595 -0.0145856   0.999839 0.00020164 0.00556244 -0.0155157   0.999864
#> 0.00020164 0.00556244 -0.0155157   0.999864 -0.00546729  0.00305068  -0.0165865    0.999843
#> -0.00546729  0.00305068  -0.0165865    0.999843  -0.0108687 0.000877602  -0.0178097    0.999782
#>  -0.0108687 0.000877602  -0.0178097    0.999782   -0.0161333 -0.000923916   -0.0191381     0.999686
#>   -0.0161333 -0.000923916   -0.0191381     0.999686  -0.0213609 -0.00233384  -0.0204926    0.999559
#>  -0.0213609 -0.00233384  -0.0204926    0.999559 -0.0266679 -0.0033953 -0.0217906   0.999401
#> -0.0266679 -0.0033953 -0.0217906   0.999401  -0.0320094 -0.00408069  -0.0229452    0.999216
#>  -0.0320094 -0.00408069  -0.0229452    0.999216  -0.0372396 -0.00438244  -0.0239579     0.99901
#>  -0.0372396 -0.00438244  -0.0239579     0.99901  -0.0423956 -0.00446108  -0.0248828    0.998781
#>  -0.0423956 -0.00446108  -0.0248828    0.998781  -0.0476911 -0.00441767  -0.0257748     0.99852
#>  -0.0476911 -0.00441767  -0.0257748     0.99852  -0.0532632 -0.00433827  -0.0266341    0.998216
#>  -0.0532632 -0.00433827  -0.0266341    0.998216  -0.0588981 -0.00420727  -0.0274287    0.997878
#>  -0.0588981 -0.00420727  -0.0274287    0.997878  -0.0644819 -0.00400327  -0.0281336    0.997514
#>  -0.0644819 -0.00400327  -0.0281336    0.997514 -0.0700113 -0.0037471 -0.0287754   0.997124
#> -0.0700113 -0.0037471 -0.0287754   0.997124  -0.0756035 -0.00351595  -0.0293978    0.996698
#>  -0.0756035 -0.00351595  -0.0293978    0.996698  -0.0813921 -0.00340295  -0.0299766    0.996225
#>  -0.0813921 -0.00340295  -0.0299766    0.996225  -0.0870973 -0.00349618  -0.0304535    0.995728
#>  -0.0870973 -0.00349618  -0.0304535    0.995728  -0.0923399 -0.00391371  -0.0308808    0.995241
#>  -0.0923399 -0.00391371  -0.0308808    0.995241 -0.0966045 -0.0046332 -0.0312726   0.994821
#> -0.0966045 -0.0046332 -0.0312726   0.994821  -0.0995227 -0.00564184  -0.0316197    0.994517
#>  -0.0995227 -0.00564184  -0.0316197    0.994517   -0.100967 -0.00677757  -0.0320055    0.994352
#>   -0.100967 -0.00677757  -0.0320055    0.994352   -0.101164 -0.00797129  -0.0324417    0.994309
#>   -0.101164 -0.00797129  -0.0324417    0.994309   -0.100512 -0.00905682  -0.0328674    0.994352
#>   -0.100512 -0.00905682  -0.0328674    0.994352 -0.0992096 -0.0100689 -0.0332089   0.994461
#> -0.0992096 -0.0100689 -0.0332089   0.994461 -0.0978375 -0.0109728 -0.0333975   0.994581
#> -0.0978375 -0.0109728 -0.0333975   0.994581 -0.0966223 -0.0118441 -0.0333687   0.994691
#> -0.0966223 -0.0118441 -0.0333687   0.994691 -0.0954228 -0.0127688 -0.0331083   0.994804
#> -0.0954228 -0.0127688 -0.0331083   0.994804 -0.0939282 -0.0140163 -0.0325646   0.994948
#> -0.0939282 -0.0140163 -0.0325646   0.994948 -0.0922052 -0.0158747  -0.031765   0.995107
#> -0.0922052 -0.0158747  -0.031765   0.995107  -0.090214 -0.0185812 -0.0306674   0.995277
#>  -0.090214 -0.0185812 -0.0306674   0.995277 -0.0880889 -0.0219422  -0.029255   0.995441
#> -0.0880889 -0.0219422  -0.029255   0.995441 -0.0858368 -0.0258901  -0.027234     0.9956
#> -0.0858368 -0.0258901  -0.027234     0.9956 -0.0835969 -0.0302973 -0.0244964   0.995738
#> -0.0835969 -0.0302973 -0.0244964   0.995738 -0.0814246 -0.0349375 -0.0209043   0.995848
#> -0.0814246 -0.0349375 -0.0209043   0.995848 -0.0793203 -0.0394466  -0.016622    0.99593
#> -0.0793203 -0.0394466  -0.016622    0.99593 -0.0772903 -0.0436046 -0.0118584   0.995984
#> -0.0772903 -0.0436046 -0.0118584   0.995984  -0.0753683  -0.0472508 -0.00698294    0.996011
#>  -0.0753683  -0.0472508 -0.00698294    0.996011  -0.0732098  -0.0503844 -0.00223732    0.996041
#>  -0.0732098  -0.0503844 -0.00223732    0.996041 -0.0708175  -0.052962 0.00200797    0.99608
#> -0.0708175  -0.052962 0.00200797    0.99608 -0.0681323 -0.0550919 0.00578812   0.996137
#> -0.0681323 -0.0550919 0.00578812   0.996137 -0.0654632 -0.0563562 0.00866892   0.996225
#> -0.0654632 -0.0563562 0.00866892   0.996225 -0.0630866 -0.0566844  0.0107263   0.996339
#> -0.0630866 -0.0566844  0.0107263   0.996339 -0.0611339 -0.0557961  0.0122109   0.996494
#> -0.0611339 -0.0557961  0.0122109   0.996494 -0.0594962 -0.0536126  0.0133537   0.996698
#> -0.0594962 -0.0536126  0.0133537   0.996698  -0.057938 -0.0501074  0.0144003   0.996958
#>  -0.057938 -0.0501074  0.0144003   0.996958 -0.0564049 -0.0453882  0.0155808   0.997254
#> -0.0564049 -0.0453882  0.0155808   0.997254 -0.0546674 -0.0396007  0.0169778   0.997575
#> -0.0546674 -0.0396007  0.0169778   0.997575 -0.0525676 -0.0329035  0.0185804   0.997902
#> -0.0525676 -0.0329035  0.0185804   0.997902 -0.0499664 -0.0254998   0.020466   0.998216
#> -0.0499664 -0.0254998   0.020466   0.998216 -0.0466811  -0.017578  0.0225805     0.9985
#> -0.0466811  -0.017578  0.0225805     0.9985  -0.0428806 -0.00931623   0.0248504    0.998728
#>  -0.0428806 -0.00931623   0.0248504    0.998728  -0.0382587 -0.00103085   0.0271669    0.998898
#>  -0.0382587 -0.00103085   0.0271669    0.998898 -0.0330564 0.00736594  0.0294892   0.998991
#> -0.0330564 0.00736594  0.0294892   0.998991 -0.0273652   0.015546  0.0316672   0.999003
#> -0.0273652   0.015546  0.0316672   0.999003 -0.0213456  0.0233696   0.033583   0.998935
#> -0.0213456  0.0233696   0.033583   0.998935 -0.0155475  0.0305347  0.0351367   0.998795
#> -0.0155475  0.0305347  0.0351367   0.998795 -0.00975784   0.0371641   0.0363614      0.9986
#> -0.00975784   0.0371641   0.0363614      0.9986 -0.00422549   0.0429545   0.0371363    0.998378
#> -0.00422549   0.0429545   0.0371363    0.998378 0.00118148  0.0480268  0.0374725   0.998142
#> 0.00118148  0.0480268  0.0374725   0.998142 0.00657714  0.0522869  0.0373488   0.997912
#> 0.00657714  0.0522869  0.0373488   0.997912 0.0122976 0.0559271 0.0367869  0.997681
#> 0.0122976 0.0559271 0.0367869  0.997681 0.0184312 0.0588753 0.0357745  0.997454
#> 0.0184312 0.0588753 0.0357745  0.997454 0.0248119 0.0610689 0.0343369  0.997234
#> 0.0248119 0.0610689 0.0343369  0.997234 0.0307635 0.0624128 0.0325422  0.997045
#> 0.0307635 0.0624128 0.0325422  0.997045 0.0363478 0.0629243 0.0304491  0.996891
#> 0.0363478 0.0629243 0.0304491  0.996891 0.0416452 0.0628909 0.0283296  0.996749
#> 0.0416452 0.0628909 0.0283296  0.996749 0.0471602 0.0626418 0.0262569  0.996575
#> 0.0471602 0.0626418 0.0262569  0.996575 0.0528048 0.0622231 0.0242957  0.996368
#> 0.0528048 0.0622231 0.0242957  0.996368 0.0584856 0.0616926 0.0224452  0.996127
#> 0.0584856 0.0616926 0.0224452  0.996127 0.0636614 0.0609984 0.0206491  0.995892
#> 0.0636614 0.0609984 0.0206491  0.995892  0.068089 0.0604629 0.0188999  0.995666
#>  0.068089 0.0604629 0.0188999  0.995666 0.0714438  0.060304 0.0172335  0.995471
#> 0.0714438  0.060304 0.0172335  0.995471 0.0737504 0.0605146 0.0156999  0.995315
#> 0.0737504 0.0605146 0.0156999  0.995315 0.0751925 0.0610659 0.0143084  0.995195
#> 0.0751925 0.0610659 0.0143084  0.995195 0.0760179 0.0621543  0.013083  0.995081
#> 0.0760179 0.0621543  0.013083  0.995081 0.0768862 0.0638074 0.0121287  0.994922
#> 0.0768862 0.0638074 0.0121287  0.994922 0.0781167 0.0660084  0.011403  0.994691
#> 0.0781167 0.0660084  0.011403  0.994691 0.0795904 0.0687124 0.0109111  0.994397
#> 0.0795904 0.0687124 0.0109111  0.994397 0.0811017 0.0713321 0.0105422  0.994094
#> 0.0811017 0.0713321 0.0105422  0.994094  0.0828199  0.0736952 0.00993852   0.993786
#> DTW alignment object
#>  Alignment size (query x reference): 101 x 101
#>  Call: dtw::dtw(x = M, step.pattern = step_pattern, distance.only = distance_only)