QTS Hemispherization

This function ensures that there are no discontinuities in QTS due to quaternion flips since two unit quaternions q and -q encode the same rotation.

Usage

hemispherize(x)

# S3 method for qts
hemispherize(x)

# S3 method for qts_sample
hemispherize(x)

Arguments

x

An object of class qts or qts_sample.

Value

An object of the same class as the input argument x with no quaternion flip discontinuities.

Examples

hemispherize(vespa64$igp[[1]])
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99427   0.07973   0.06988   0.01334
#>  2     1   0.99483   0.07457   0.06763   0.01313
#>  3     2   0.99542   0.06931   0.06457   0.01269
#>  4     3   0.99602   0.06398   0.06091   0.01184
#>  5     4   0.99652   0.05949   0.05742   0.01070
#>  6     5   0.99694   0.05572   0.05403   0.00932
#>  7     6   0.99729   0.05275   0.05079   0.00772
#>  8     7   0.99757   0.05056   0.04761   0.00583
#>  9     8   0.99782   0.04883   0.04430   0.00366
#> 10     9   0.99805   0.04730   0.04071   0.00125
#> # … with 91 more rows
hemispherize(vespa64$igp)
#> [[1]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99427   0.07973   0.06988   0.01334
#>  2     1   0.99483   0.07457   0.06763   0.01313
#>  3     2   0.99542   0.06931   0.06457   0.01269
#>  4     3   0.99602   0.06398   0.06091   0.01184
#>  5     4   0.99652   0.05949   0.05742   0.01070
#>  6     5   0.99694   0.05572   0.05403   0.00932
#>  7     6   0.99729   0.05275   0.05079   0.00772
#>  8     7   0.99757   0.05056   0.04761   0.00583
#>  9     8   0.99782   0.04883   0.04430   0.00366
#> 10     9   0.99805   0.04730   0.04071   0.00125
#> # … with 91 more rows
#> 
#> [[2]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99379   0.08282   0.07370   0.00994
#>  2     1   0.99441   0.07727   0.07120   0.01036
#>  3     2   0.99503   0.07171   0.06829   0.01059
#>  4     3   0.99565   0.06598   0.06489   0.01047
#>  5     4   0.99625   0.06032   0.06124   0.01005
#>  6     5   0.99677   0.05525   0.05756   0.00927
#>  7     6   0.99717   0.05142   0.05417   0.00810
#>  8     7   0.99751   0.04849   0.05080   0.00655
#>  9     8   0.99780   0.04622   0.04723   0.00461
#> 10     9   0.99806   0.04442   0.04345   0.00237
#> # … with 91 more rows
#> 
#> [[3]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99262   0.09948   0.06271  -0.02971
#>  2     1   0.99312   0.09388   0.06347  -0.02941
#>  3     2   0.99375   0.08743   0.06335  -0.02831
#>  4     3   0.99446   0.08006   0.06255  -0.02696
#>  5     4   0.99511   0.07312   0.06135  -0.02551
#>  6     5   0.99566   0.06703   0.05977  -0.02423
#>  7     6   0.99614   0.06183   0.05782  -0.02305
#>  8     7   0.99656   0.05763   0.05543  -0.02198
#>  9     8   0.99692   0.05426   0.05257  -0.02103
#> 10     9   0.99726   0.05151   0.04915  -0.02020
#> # … with 91 more rows
#> 
#> [[4]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99281   0.09669   0.06308  -0.03165
#>  2     1   0.99331   0.09099   0.06403  -0.03082
#>  3     2   0.99390   0.08485   0.06391  -0.02969
#>  4     3   0.99452   0.07821   0.06324  -0.02837
#>  5     4   0.99514   0.07160   0.06207  -0.02695
#>  6     5   0.99567   0.06570   0.06052  -0.02564
#>  7     6   0.99613   0.06075   0.05862  -0.02447
#>  8     7   0.99651   0.05693   0.05633  -0.02347
#>  9     8   0.99685   0.05402   0.05358  -0.02258
#> 10     9   0.99716   0.05187   0.05014  -0.02182
#> # … with 91 more rows
#> 
#> [[5]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99414   0.08053   0.07026   0.01615
#>  2     1   0.99469   0.07558   0.06822   0.01535
#>  3     2   0.99521   0.07085   0.06566   0.01506
#>  4     3   0.99574   0.06628   0.06246   0.01461
#>  5     4   0.99623   0.06202   0.05899   0.01379
#>  6     5   0.99669   0.05810   0.05541   0.01259
#>  7     6   0.99710   0.05466   0.05184   0.01102
#>  8     7   0.99743   0.05196   0.04844   0.00912
#>  9     8   0.99771   0.04994   0.04508   0.00691
#> 10     9   0.99795   0.04836   0.04157   0.00448
#> # … with 91 more rows
#> 
#> [[6]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99326   0.08804   0.07386   0.01483
#>  2     1   0.99376   0.08244   0.07368   0.01449
#>  3     2   0.99434   0.07693   0.07184   0.01450
#>  4     3   0.99498   0.07123   0.06879   0.01445
#>  5     4   0.99561   0.06567   0.06516   0.01412
#>  6     5   0.99617   0.06071   0.06152   0.01353
#>  7     6   0.99664   0.05649   0.05800   0.01258
#>  8     7   0.99700   0.05351   0.05475   0.01112
#>  9     8   0.99730   0.05149   0.05157   0.00922
#> 10     9   0.99757   0.04996   0.04815   0.00687
#> # … with 91 more rows
#> 
#> [[7]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99544   0.07361   0.03614  -0.04869
#>  2     1   0.99563   0.07000   0.03835  -0.04843
#>  3     2   0.99579   0.06682   0.04066  -0.04779
#>  4     3   0.99597   0.06369   0.04263  -0.04663
#>  5     4   0.99617   0.06045   0.04409  -0.04516
#>  6     5   0.99640   0.05724   0.04502  -0.04345
#>  7     6   0.99663   0.05414   0.04557  -0.04139
#>  8     7   0.99688   0.05142   0.04526  -0.03931
#>  9     8   0.99712   0.04916   0.04413  -0.03727
#> 10     9   0.99736   0.04716   0.04244  -0.03530
#> # … with 91 more rows
#> 
#> [[8]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99506   0.06909   0.04686  -0.05375
#>  2     1   0.99525   0.06662   0.04888  -0.05157
#>  3     2   0.99547   0.06355   0.05103  -0.04893
#>  4     3   0.99572   0.06026   0.05258  -0.04627
#>  5     4   0.99599   0.05693   0.05343  -0.04370
#>  6     5   0.99627   0.05376   0.05346  -0.04124
#>  7     6   0.99656   0.05103   0.05246  -0.03897
#>  8     7   0.99685   0.04876   0.05047  -0.03688
#>  9     8   0.99715   0.04701   0.04758  -0.03491
#> 10     9   0.99745   0.04557   0.04389  -0.03303
#> # … with 91 more rows
#> 
#> [[9]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99416   0.08373   0.06392   0.02332
#>  2     1   0.99469   0.07722   0.06345   0.02448
#>  3     2   0.99533   0.06977   0.06172   0.02525
#>  4     3   0.99605   0.06144   0.05870   0.02562
#>  5     4   0.99673   0.05348   0.05494   0.02533
#>  6     5   0.99733   0.04646   0.05084   0.02434
#>  7     6   0.99780   0.04115   0.04679   0.02247
#>  8     7   0.99816   0.03771   0.04320   0.01981
#>  9     8   0.99842   0.03582   0.04020   0.01636
#> 10     9   0.99859   0.03487   0.03794   0.01247
#> # … with 91 more rows
#> 
#> [[10]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99410   0.08159   0.06763   0.02292
#>  2     1   0.99466   0.07394   0.06792   0.02377
#>  3     2   0.99529   0.06609   0.06654   0.02472
#>  4     3   0.99595   0.05819   0.06373   0.02520
#>  5     4   0.99660   0.05040   0.06008   0.02510
#>  6     5   0.99718   0.04344   0.05611   0.02428
#>  7     6   0.99766   0.03784   0.05219   0.02265
#>  8     7   0.99802   0.03411   0.04873   0.02029
#>  9     8   0.99829   0.03204   0.04579   0.01737
#> 10     9   0.99846   0.03142   0.04348   0.01395
#> # … with 91 more rows
#> 
#> [[11]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99378   0.09127   0.06209   0.01488
#>  2     1   0.99427   0.08384   0.06456   0.01546
#>  3     2   0.99474   0.07662   0.06608   0.01613
#>  4     3   0.99520   0.07008   0.06629   0.01666
#>  5     4   0.99560   0.06501   0.06528   0.01686
#>  6     5   0.99600   0.06116   0.06301   0.01667
#>  7     6   0.99642   0.05817   0.05930   0.01604
#>  8     7   0.99687   0.05533   0.05442   0.01509
#>  9     8   0.99731   0.05265   0.04899   0.01382
#> 10     9   0.99771   0.05023   0.04370   0.01222
#> # … with 91 more rows
#> 
#> [[12]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99378   0.09099   0.06243   0.01470
#>  2     1   0.99423   0.08331   0.06576   0.01532
#>  3     2   0.99467   0.07589   0.06787   0.01635
#>  4     3   0.99510   0.06927   0.06838   0.01739
#>  5     4   0.99556   0.06364   0.06695   0.01831
#>  6     5   0.99606   0.05855   0.06378   0.01903
#>  7     6   0.99659   0.05407   0.05925   0.01932
#>  8     7   0.99710   0.05016   0.05397   0.01914
#>  9     8   0.99753   0.04716   0.04864   0.01831
#> 10     9   0.99786   0.04505   0.04430   0.01688
#> # … with 91 more rows
#> 
#> [[13]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99572   0.05082   0.07120   0.02989
#>  2     1   0.99607   0.04731   0.06851   0.03020
#>  3     2   0.99652   0.04346   0.06424   0.03059
#>  4     3   0.99690   0.04018   0.06048   0.03024
#>  5     4   0.99721   0.03769   0.05743   0.02918
#>  6     5   0.99745   0.03611   0.05520   0.02739
#>  7     6   0.99762   0.03522   0.05363   0.02507
#>  8     7   0.99777   0.03469   0.05249   0.02237
#>  9     8   0.99790   0.03413   0.05149   0.01936
#> 10     9   0.99805   0.03312   0.05041   0.01624
#> # … with 91 more rows
#> 
#> [[14]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99533   0.05730   0.07186   0.02939
#>  2     1   0.99564   0.05383   0.06994   0.03028
#>  3     2   0.99602   0.04977   0.06725   0.03076
#>  4     3   0.99642   0.04603   0.06398   0.03065
#>  5     4   0.99679   0.04294   0.06059   0.02990
#>  6     5   0.99711   0.04071   0.05741   0.02849
#>  7     6   0.99739   0.03935   0.05452   0.02648
#>  8     7   0.99761   0.03863   0.05191   0.02407
#>  9     8   0.99782   0.03810   0.04948   0.02144
#> 10     9   0.99802   0.03731   0.04712   0.01868
#> # … with 91 more rows
#> 
#> [[15]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99572   0.04763   0.07857   0.00966
#>  2     1   0.99606   0.04320   0.07662   0.01099
#>  3     2   0.99645   0.03916   0.07353   0.01212
#>  4     3   0.99685   0.03575   0.06964   0.01263
#>  5     4   0.99723   0.03312   0.06537   0.01245
#>  6     5   0.99757   0.03146   0.06115   0.01154
#>  7     6   0.99784   0.03067   0.05716   0.00996
#>  8     7   0.99807   0.03046   0.05351   0.00791
#>  9     8   0.99826   0.03047   0.05026   0.00564
#> 10     9   0.99841   0.03027   0.04742   0.00342
#> # … with 91 more rows
#> 
#> [[16]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99508   0.05950   0.07884   0.00834
#>  2     1   0.99543   0.05417   0.07795   0.00998
#>  3     2   0.99583   0.04959   0.07573   0.01138
#>  4     3   0.99625   0.04556   0.07256   0.01228
#>  5     4   0.99665   0.04230   0.06889   0.01263
#>  6     5   0.99701   0.03996   0.06501   0.01235
#>  7     6   0.99731   0.03857   0.06124   0.01144
#>  8     7   0.99757   0.03784   0.05758   0.00995
#>  9     8   0.99781   0.03740   0.05400   0.00803
#> 10     9   0.99803   0.03681   0.05039   0.00590
#> # … with 91 more rows
#> 
#> [[17]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99629   0.06356   0.03543   0.04602
#>  2     1   0.99660   0.05922   0.03558   0.04486
#>  3     2   0.99691   0.05521   0.03517   0.04352
#>  4     3   0.99720   0.05120   0.03477   0.04191
#>  5     4   0.99748   0.04739   0.03443   0.04010
#>  6     5   0.99773   0.04383   0.03425   0.03798
#>  7     6   0.99795   0.04065   0.03422   0.03554
#>  8     7   0.99816   0.03779   0.03427   0.03282
#>  9     8   0.99834   0.03510   0.03437   0.02992
#> 10     9   0.99852   0.03247   0.03447   0.02693
#> # … with 91 more rows
#> 
#> [[18]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99610   0.06296   0.03600   0.05024
#>  2     1   0.99647   0.05819   0.03511   0.04922
#>  3     2   0.99682   0.05376   0.03413   0.04794
#>  4     3   0.99714   0.04963   0.03342   0.04628
#>  5     4   0.99742   0.04591   0.03294   0.04434
#>  6     5   0.99768   0.04262   0.03259   0.04202
#>  7     6   0.99791   0.03983   0.03232   0.03935
#>  8     7   0.99811   0.03752   0.03219   0.03639
#>  9     8   0.99831   0.03524   0.03216   0.03314
#> 10     9   0.99849   0.03297   0.03211   0.02984
#> # … with 91 more rows
#> 
#> [[19]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99425   0.08755   0.06114   0.00736
#>  2     1   0.99449   0.08382   0.06244   0.00782
#>  3     2   0.99479   0.08000   0.06257   0.00848
#>  4     3   0.99520   0.07567   0.06144   0.00925
#>  5     4   0.99564   0.07109   0.05950   0.00998
#>  6     5   0.99611   0.06640   0.05694   0.01054
#>  7     6   0.99657   0.06187   0.05394   0.01082
#>  8     7   0.99701   0.05748   0.05054   0.01064
#>  9     8   0.99743   0.05342   0.04676   0.00993
#> 10     9   0.99780   0.04976   0.04281   0.00882
#> # … with 91 more rows
#> 
#> [[20]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99415   0.07971   0.07196   0.01186
#>  2     1   0.99466   0.07439   0.07042   0.01248
#>  3     2   0.99519   0.06930   0.06793   0.01330
#>  4     3   0.99574   0.06414   0.06482   0.01376
#>  5     4   0.99624   0.05939   0.06150   0.01371
#>  6     5   0.99668   0.05542   0.05816   0.01313
#>  7     6   0.99707   0.05212   0.05470   0.01209
#>  8     7   0.99744   0.04914   0.05093   0.01066
#>  9     8   0.99779   0.04641   0.04673   0.00885
#> 10     9   0.99812   0.04383   0.04233   0.00685
#> # … with 91 more rows
#> 
#> [[21]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99514   0.07774   0.04211   0.04334
#>  2     1   0.99562   0.07217   0.04124   0.04284
#>  3     2   0.99604   0.06719   0.04020   0.04223
#>  4     3   0.99643   0.06235   0.03934   0.04121
#>  5     4   0.99677   0.05797   0.03869   0.03978
#>  6     5   0.99709   0.05391   0.03829   0.03782
#>  7     6   0.99739   0.05013   0.03808   0.03535
#>  8     7   0.99767   0.04652   0.03797   0.03254
#>  9     8   0.99791   0.04310   0.03794   0.02960
#> 10     9   0.99814   0.03966   0.03796   0.02657
#> # … with 91 more rows
#> 
#> [[22]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99565   0.07073   0.04537   0.04024
#>  2     1   0.99606   0.06590   0.04397   0.03998
#>  3     2   0.99647   0.06078   0.04244   0.03931
#>  4     3   0.99686   0.05577   0.04130   0.03818
#>  5     4   0.99718   0.05137   0.04071   0.03660
#>  6     5   0.99745   0.04760   0.04043   0.03458
#>  7     6   0.99769   0.04431   0.04019   0.03208
#>  8     7   0.99790   0.04154   0.04000   0.02935
#>  9     8   0.99810   0.03893   0.03983   0.02645
#> 10     9   0.99829   0.03604   0.03958   0.02341
#> # … with 91 more rows
#> 
#> [[23]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99396   0.08718   0.06504   0.01470
#>  2     1   0.99426   0.08295   0.06581   0.01518
#>  3     2   0.99460   0.07898   0.06531   0.01616
#>  4     3   0.99502   0.07470   0.06373   0.01723
#>  5     4   0.99549   0.07014   0.06126   0.01814
#>  6     5   0.99597   0.06555   0.05823   0.01865
#>  7     6   0.99646   0.06107   0.05481   0.01854
#>  8     7   0.99691   0.05687   0.05120   0.01790
#>  9     8   0.99734   0.05292   0.04730   0.01672
#> 10     9   0.99775   0.04922   0.04309   0.01499
#> # … with 91 more rows
#> 
#> [[24]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99353   0.08626   0.07223   0.01578
#>  2     1   0.99390   0.08166   0.07244   0.01573
#>  3     2   0.99423   0.07778   0.07201   0.01643
#>  4     3   0.99463   0.07372   0.07052   0.01716
#>  5     4   0.99511   0.06934   0.06808   0.01786
#>  6     5   0.99561   0.06483   0.06492   0.01833
#>  7     6   0.99613   0.06035   0.06121   0.01838
#>  8     7   0.99662   0.05618   0.05723   0.01789
#>  9     8   0.99709   0.05230   0.05292   0.01687
#> 10     9   0.99753   0.04868   0.04831   0.01535
#> # … with 91 more rows
#> 
#> [[25]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99682   0.06287   0.04827   0.00874
#>  2     1   0.99698   0.06210   0.04587   0.00788
#>  3     2   0.99728   0.05994   0.04232   0.00711
#>  4     3   0.99762   0.05693   0.03836   0.00642
#>  5     4   0.99799   0.05310   0.03406   0.00576
#>  6     5   0.99838   0.04835   0.02945   0.00511
#>  7     6   0.99879   0.04245   0.02453   0.00458
#>  8     7   0.99915   0.03587   0.01970   0.00403
#>  9     8   0.99947   0.02881   0.01491   0.00350
#> 10     9   0.99971   0.02142   0.01022   0.00297
#> # … with 91 more rows
#> 
#> [[26]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99508   0.06775   0.07115   0.01238
#>  2     1   0.99555   0.06453   0.06735   0.01331
#>  3     2   0.99601   0.06114   0.06356   0.01385
#>  4     3   0.99643   0.05771   0.06000   0.01413
#>  5     4   0.99683   0.05401   0.05668   0.01412
#>  6     5   0.99716   0.05091   0.05380   0.01387
#>  7     6   0.99741   0.04853   0.05133   0.01328
#>  8     7   0.99764   0.04651   0.04886   0.01243
#>  9     8   0.99786   0.04483   0.04616   0.01134
#> 10     9   0.99808   0.04324   0.04329   0.01008
#> # … with 91 more rows
#> 
#> [[27]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99730   0.02816   0.06431  -0.02133
#>  2     1   0.99758   0.02689   0.06127  -0.01878
#>  3     2   0.99785   0.02634   0.05769  -0.01654
#>  4     3   0.99806   0.02657   0.05419  -0.01502
#>  5     4   0.99824   0.02739   0.05069  -0.01393
#>  6     5   0.99838   0.02871   0.04734  -0.01332
#>  7     6   0.99848   0.03007   0.04429  -0.01323
#>  8     7   0.99857   0.03112   0.04127  -0.01344
#>  9     8   0.99867   0.03179   0.03807  -0.01381
#> 10     9   0.99879   0.03198   0.03461  -0.01417
#> # … with 91 more rows
#> 
#> [[28]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99641   0.03998   0.06902  -0.02850
#>  2     1   0.99670   0.03666   0.06738  -0.02647
#>  3     2   0.99697   0.03464   0.06528  -0.02421
#>  4     3   0.99722   0.03345   0.06290  -0.02203
#>  5     4   0.99741   0.03316   0.06057  -0.02016
#>  6     5   0.99755   0.03374   0.05831  -0.01882
#>  7     6   0.99767   0.03488   0.05587  -0.01791
#>  8     7   0.99777   0.03627   0.05325  -0.01754
#>  9     8   0.99788   0.03741   0.05029  -0.01767
#> 10     9   0.99800   0.03815   0.04695  -0.01810
#> # … with 91 more rows
#> 
#> [[29]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99647   0.03632   0.07527  -0.00825
#>  2     1   0.99677   0.03462   0.07232  -0.00560
#>  3     2   0.99709   0.03272   0.06879  -0.00297
#>  4     3   0.99736   0.03145   0.06544  -0.00080
#>  5     4   0.99756   0.03084   0.06268   0.00070
#>  6     5   0.99768   0.03090   0.06060   0.00141
#>  7     6   0.99778   0.03135   0.05872   0.00149
#>  8     7   0.99787   0.03194   0.05679   0.00109
#>  9     8   0.99799   0.03244   0.05446   0.00039
#> 10     9   0.99814   0.03262   0.05155  -0.00048
#> # … with 91 more rows
#> 
#> [[30]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99652   0.03337   0.07618  -0.00600
#>  2     1   0.99673   0.03330   0.07345  -0.00403
#>  3     2   0.99697   0.03191   0.07091  -0.00209
#>  4     3   0.99718   0.03104   0.06827  -0.00061
#>  5     4   0.99735   0.03081   0.06584   0.00031
#>  6     5   0.99750   0.03111   0.06350   0.00068
#>  7     6   0.99762   0.03176   0.06119   0.00052
#>  8     7   0.99775   0.03250   0.05865   0.00002
#>  9     8   0.99790   0.03305   0.05570  -0.00068
#> 10     9   0.99808   0.03324   0.05215  -0.00142
#> # … with 91 more rows
#> 
#> [[31]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99695   0.02603   0.06462  -0.03509
#>  2     1   0.99715   0.02464   0.06372  -0.03205
#>  3     2   0.99733   0.02390   0.06252  -0.02931
#>  4     3   0.99750   0.02361   0.06090  -0.02701
#>  5     4   0.99765   0.02390   0.05900  -0.02531
#>  6     5   0.99778   0.02464   0.05692  -0.02412
#>  7     6   0.99792   0.02553   0.05453  -0.02317
#>  8     7   0.99806   0.02641   0.05176  -0.02238
#>  9     8   0.99823   0.02702   0.04845  -0.02163
#> 10     9   0.99841   0.02730   0.04462  -0.02077
#> # … with 91 more rows
#> 
#> [[32]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99718   0.02189   0.06453  -0.03153
#>  2     1   0.99738   0.02053   0.06279  -0.02948
#>  3     2   0.99754   0.02023   0.06113  -0.02756
#>  4     3   0.99768   0.02043   0.05944  -0.02599
#>  5     4   0.99780   0.02124   0.05757  -0.02489
#>  6     5   0.99793   0.02243   0.05530  -0.02414
#>  7     6   0.99806   0.02383   0.05251  -0.02357
#>  8     7   0.99821   0.02526   0.04915  -0.02303
#>  9     8   0.99837   0.02651   0.04518  -0.02245
#> 10     9   0.99856   0.02732   0.04071  -0.02168
#> # … with 91 more rows
#> 
#> [[33]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99099   0.12441   0.00926   0.04863
#>  2     1   0.99155   0.11940   0.01256   0.04922
#>  3     2   0.99205   0.11457   0.01605   0.04963
#>  4     3   0.99252   0.10988   0.01929   0.04958
#>  5     4   0.99294   0.10561   0.02232   0.04913
#>  6     5   0.99330   0.10189   0.02525   0.04829
#>  7     6   0.99362   0.09861   0.02802   0.04694
#>  8     7   0.99393   0.09555   0.03055   0.04509
#>  9     8   0.99425   0.09249   0.03282   0.04274
#> 10     9   0.99459   0.08940   0.03481   0.03993
#> # … with 91 more rows
#> 
#> [[34]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99277   0.11089   0.00761   0.04529
#>  2     1   0.99300   0.10838   0.01070   0.04572
#>  3     2   0.99333   0.10515   0.01362   0.04536
#>  4     3   0.99364   0.10217   0.01647   0.04446
#>  5     4   0.99391   0.09970   0.01907   0.04299
#>  6     5   0.99417   0.09737   0.02145   0.04102
#>  7     6   0.99446   0.09491   0.02365   0.03855
#>  8     7   0.99478   0.09219   0.02540   0.03562
#>  9     8   0.99513   0.08920   0.02668   0.03246
#> 10     9   0.99549   0.08596   0.02762   0.02908
#> # … with 91 more rows
#> 
#> [[35]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99036   0.12354   0.04394   0.04466
#>  2     1   0.99088   0.11901   0.04565   0.04363
#>  3     2   0.99143   0.11455   0.04681   0.04199
#>  4     3   0.99196   0.11035   0.04732   0.04010
#>  5     4   0.99243   0.10677   0.04729   0.03797
#>  6     5   0.99284   0.10402   0.04663   0.03564
#>  7     6   0.99325   0.10157   0.04532   0.03298
#>  8     7   0.99363   0.09944   0.04347   0.03021
#>  9     8   0.99404   0.09730   0.04089   0.02717
#> 10     9   0.99449   0.09499   0.03749   0.02380
#> # … with 91 more rows
#> 
#> [[36]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99024   0.12061   0.05259   0.04604
#>  2     1   0.99083   0.11574   0.05404   0.04411
#>  3     2   0.99142   0.11085   0.05503   0.04203
#>  4     3   0.99199   0.10647   0.05502   0.03983
#>  5     4   0.99252   0.10276   0.05411   0.03755
#>  6     5   0.99304   0.09959   0.05221   0.03506
#>  7     6   0.99357   0.09664   0.04938   0.03227
#>  8     7   0.99410   0.09393   0.04576   0.02914
#>  9     8   0.99462   0.09129   0.04157   0.02581
#> 10     9   0.99512   0.08868   0.03691   0.02237
#> # … with 91 more rows
#> 
#> [[37]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99204   0.11879   0.00199   0.04178
#>  2     1   0.99243   0.11459   0.00592   0.04382
#>  3     2   0.99283   0.10998   0.00986   0.04569
#>  4     3   0.99320   0.10567   0.01385   0.04688
#>  5     4   0.99347   0.10229   0.01754   0.04733
#>  6     5   0.99367   0.09967   0.02114   0.04723
#>  7     6   0.99385   0.09742   0.02461   0.04655
#>  8     7   0.99404   0.09523   0.02795   0.04520
#>  9     8   0.99426   0.09281   0.03113   0.04319
#> 10     9   0.99451   0.09018   0.03398   0.04067
#> # … with 91 more rows
#> 
#> [[38]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99236   0.11709   0.00972   0.03775
#>  2     1   0.99277   0.11272   0.01370   0.03890
#>  3     2   0.99317   0.10845   0.01719   0.03952
#>  4     3   0.99355   0.10441   0.02012   0.03948
#>  5     4   0.99390   0.10057   0.02292   0.03890
#>  6     5   0.99426   0.09675   0.02561   0.03773
#>  7     6   0.99463   0.09290   0.02805   0.03610
#>  8     7   0.99497   0.08923   0.03020   0.03406
#>  9     8   0.99532   0.08554   0.03209   0.03156
#> 10     9   0.99568   0.08167   0.03377   0.02858
#> # … with 91 more rows
#> 
#> [[39]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99012   0.12860   0.02482   0.05007
#>  2     1   0.99039   0.12633   0.02801   0.04892
#>  3     2   0.99081   0.12282   0.03027   0.04794
#>  4     3   0.99127   0.11920   0.03170   0.04655
#>  5     4   0.99173   0.11584   0.03223   0.04477
#>  6     5   0.99220   0.11279   0.03179   0.04255
#>  7     6   0.99266   0.10993   0.03071   0.03997
#>  8     7   0.99314   0.10702   0.02903   0.03700
#>  9     8   0.99365   0.10394   0.02678   0.03362
#> 10     9   0.99415   0.10086   0.02419   0.03015
#> # … with 91 more rows
#> 
#> [[40]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.98906   0.13524   0.03494   0.04731
#>  2     1   0.98966   0.12986   0.03961   0.04629
#>  3     2   0.99021   0.12502   0.04267   0.04506
#>  4     3   0.99071   0.12091   0.04445   0.04359
#>  5     4   0.99117   0.11750   0.04506   0.04186
#>  6     5   0.99162   0.11461   0.04450   0.03976
#>  7     6   0.99208   0.11207   0.04278   0.03727
#>  8     7   0.99256   0.10972   0.04007   0.03441
#>  9     8   0.99307   0.10726   0.03646   0.03117
#> 10     9   0.99361   0.10463   0.03217   0.02767
#> # … with 91 more rows
#> 
#> [[41]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99204   0.11046   0.04906   0.03525
#>  2     1   0.99232   0.10801   0.05000   0.03377
#>  3     2   0.99267   0.10519   0.05021   0.03204
#>  4     3   0.99306   0.10238   0.04944   0.03009
#>  5     4   0.99348   0.09980   0.04765   0.02782
#>  6     5   0.99390   0.09752   0.04493   0.02527
#>  7     6   0.99432   0.09534   0.04169   0.02242
#>  8     7   0.99474   0.09315   0.03815   0.01923
#>  9     8   0.99514   0.09082   0.03446   0.01589
#> 10     9   0.99557   0.08809   0.03059   0.01242
#> # … with 91 more rows
#> 
#> [[42]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99155   0.11826   0.04717   0.02475
#>  2     1   0.99222   0.11290   0.04696   0.02342
#>  3     2   0.99279   0.10856   0.04592   0.02187
#>  4     3   0.99326   0.10515   0.04445   0.02007
#>  5     4   0.99365   0.10255   0.04271   0.01806
#>  6     5   0.99398   0.10046   0.04069   0.01583
#>  7     6   0.99431   0.09848   0.03839   0.01338
#>  8     7   0.99462   0.09654   0.03586   0.01080
#>  9     8   0.99496   0.09434   0.03296   0.00820
#> 10     9   0.99535   0.09147   0.02959   0.00546
#> # … with 91 more rows
#> 
#> [[43]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.98866   0.13122   0.07173   0.01356
#>  2     1   0.98959   0.12405   0.07199   0.01163
#>  3     2   0.99045   0.11733   0.07173   0.00996
#>  4     3   0.99120   0.11143   0.07090   0.00855
#>  5     4   0.99183   0.10664   0.06957   0.00726
#>  6     5   0.99238   0.10277   0.06771   0.00602
#>  7     6   0.99286   0.09964   0.06540   0.00482
#>  8     7   0.99329   0.09713   0.06259   0.00366
#>  9     8   0.99373   0.09485   0.05906   0.00250
#> 10     9   0.99420   0.09252   0.05479   0.00137
#> # … with 91 more rows
#> 
#> [[44]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.98965   0.12271   0.07318   0.01328
#>  2     1   0.99052   0.11563   0.07327   0.01111
#>  3     2   0.99116   0.11041   0.07295   0.00959
#>  4     3   0.99165   0.10655   0.07222   0.00833
#>  5     4   0.99205   0.10369   0.07096   0.00716
#>  6     5   0.99245   0.10120   0.06908   0.00604
#>  7     6   0.99287   0.09887   0.06639   0.00503
#>  8     7   0.99332   0.09656   0.06303   0.00418
#>  9     8   0.99379   0.09421   0.05906   0.00353
#> 10     9   0.99428   0.09171   0.05464   0.00309
#> # … with 91 more rows
#> 
#> [[45]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99199   0.11271   0.04789   0.03094
#>  2     1   0.99246   0.10784   0.05006   0.02971
#>  3     2   0.99295   0.10290   0.05157   0.02829
#>  4     3   0.99339   0.09859   0.05238   0.02666
#>  5     4   0.99378   0.09501   0.05253   0.02481
#>  6     5   0.99411   0.09220   0.05218   0.02265
#>  7     6   0.99441   0.08996   0.05141   0.02028
#>  8     7   0.99469   0.08796   0.05030   0.01777
#>  9     8   0.99497   0.08618   0.04877   0.01517
#> 10     9   0.99525   0.08446   0.04684   0.01253
#> # … with 91 more rows
#> 
#> [[46]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99267   0.11118   0.03922   0.02645
#>  2     1   0.99306   0.10684   0.04257   0.02470
#>  3     2   0.99352   0.10159   0.04547   0.02293
#>  4     3   0.99394   0.09677   0.04753   0.02125
#>  5     4   0.99435   0.09229   0.04865   0.01946
#>  6     5   0.99471   0.08853   0.04913   0.01754
#>  7     6   0.99501   0.08546   0.04923   0.01534
#>  8     7   0.99525   0.08318   0.04887   0.01282
#>  9     8   0.99546   0.08155   0.04812   0.01014
#> 10     9   0.99564   0.08017   0.04704   0.00716
#> # … with 91 more rows
#> 
#> [[47]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99154   0.11263   0.06018   0.02322
#>  2     1   0.99213   0.10680   0.06168   0.02185
#>  3     2   0.99275   0.10035   0.06299   0.02007
#>  4     3   0.99330   0.09507   0.06318   0.01822
#>  5     4   0.99374   0.09100   0.06268   0.01623
#>  6     5   0.99410   0.08805   0.06167   0.01415
#>  7     6   0.99443   0.08568   0.06019   0.01200
#>  8     7   0.99475   0.08361   0.05812   0.00980
#>  9     8   0.99507   0.08187   0.05542   0.00769
#> 10     9   0.99538   0.08030   0.05227   0.00575
#> # … with 91 more rows
#> 
#> [[48]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99325   0.10171   0.05067   0.02306
#>  2     1   0.99365   0.09692   0.05324   0.02081
#>  3     2   0.99399   0.09243   0.05562   0.01889
#>  4     3   0.99429   0.08863   0.05690   0.01715
#>  5     4   0.99457   0.08556   0.05717   0.01554
#>  6     5   0.99482   0.08317   0.05672   0.01405
#>  7     6   0.99505   0.08132   0.05572   0.01267
#>  8     7   0.99525   0.08014   0.05410   0.01134
#>  9     8   0.99543   0.07944   0.05207   0.01004
#> 10     9   0.99561   0.07886   0.04972   0.00877
#> # … with 91 more rows
#> 
#> [[49]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.98737   0.14934   0.03624   0.03864
#>  2     1   0.98838   0.14207   0.03857   0.03775
#>  3     2   0.98949   0.13415   0.03981   0.03652
#>  4     3   0.99043   0.12750   0.03950   0.03505
#>  5     4   0.99117   0.12255   0.03821   0.03323
#>  6     5   0.99178   0.11881   0.03621   0.03088
#>  7     6   0.99227   0.11620   0.03354   0.02783
#>  8     7   0.99270   0.11423   0.03039   0.02419
#>  9     8   0.99306   0.11268   0.02701   0.02023
#> 10     9   0.99335   0.11150   0.02351   0.01626
#> # … with 91 more rows
#> 
#> [[50]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.98811   0.14431   0.02976   0.04386
#>  2     1   0.98906   0.13764   0.03178   0.04257
#>  3     2   0.98991   0.13174   0.03252   0.04086
#>  4     3   0.99076   0.12596   0.03199   0.03868
#>  5     4   0.99153   0.12094   0.03063   0.03598
#>  6     5   0.99213   0.11727   0.02883   0.03304
#>  7     6   0.99259   0.11473   0.02663   0.02993
#>  8     7   0.99300   0.11261   0.02392   0.02647
#>  9     8   0.99338   0.11075   0.02052   0.02253
#> 10     9   0.99373   0.10903   0.01658   0.01844
#> # … with 91 more rows
#> 
#> [[51]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.98550   0.15687   0.06393   0.00994
#>  2     1   0.98652   0.15017   0.06445   0.00799
#>  3     2   0.98771   0.14206   0.06484   0.00634
#>  4     3   0.98884   0.13402   0.06480   0.00499
#>  5     4   0.98981   0.12700   0.06426   0.00383
#>  6     5   0.99061   0.12120   0.06319   0.00272
#>  7     6   0.99127   0.11664   0.06151   0.00160
#>  8     7   0.99181   0.11323   0.05912   0.00049
#>  9     8   0.99230   0.11060   0.05579  -0.00067
#> 10     9   0.99275   0.10846   0.05183  -0.00184
#> # … with 91 more rows
#> 
#> [[52]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.98449   0.15732   0.07648   0.01350
#>  2     1   0.98559   0.15008   0.07716   0.01160
#>  3     2   0.98675   0.14217   0.07751   0.00977
#>  4     3   0.98789   0.13411   0.07759   0.00829
#>  5     4   0.98893   0.12659   0.07708   0.00681
#>  6     5   0.98983   0.12033   0.07576   0.00518
#>  7     6   0.99059   0.11530   0.07361   0.00336
#>  8     7   0.99124   0.11148   0.07084   0.00162
#>  9     8   0.99177   0.10875   0.06762   0.00008
#> 10     9   0.99226   0.10644   0.06388  -0.00142
#> # … with 91 more rows
#> 
#> [[53]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.98810   0.14740   0.03281   0.02919
#>  2     1   0.98877   0.14203   0.03655   0.02868
#>  3     2   0.98959   0.13582   0.03862   0.02770
#>  4     3   0.99039   0.13005   0.03907   0.02637
#>  5     4   0.99108   0.12527   0.03818   0.02455
#>  6     5   0.99166   0.12156   0.03645   0.02234
#>  7     6   0.99218   0.11850   0.03399   0.01970
#>  8     7   0.99262   0.11608   0.03085   0.01656
#>  9     8   0.99301   0.11417   0.02714   0.01298
#> 10     9   0.99334   0.11246   0.02319   0.00922
#> # … with 91 more rows
#> 
#> [[54]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.98740   0.14936   0.03615   0.03778
#>  2     1   0.98811   0.14414   0.03889   0.03684
#>  3     2   0.98897   0.13787   0.04071   0.03570
#>  4     3   0.98977   0.13238   0.04087   0.03420
#>  5     4   0.99055   0.12728   0.03959   0.03218
#>  6     5   0.99124   0.12313   0.03752   0.02972
#>  7     6   0.99178   0.12012   0.03493   0.02690
#>  8     7   0.99222   0.11806   0.03188   0.02354
#>  9     8   0.99260   0.11644   0.02823   0.01960
#> 10     9   0.99294   0.11508   0.02422   0.01521
#> # … with 91 more rows
#> 
#> [[55]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.98627   0.15693   0.05108   0.00545
#>  2     1   0.98723   0.14978   0.05410   0.00400
#>  3     2   0.98826   0.14207   0.05613   0.00326
#>  4     3   0.98917   0.13494   0.05760   0.00280
#>  5     4   0.98995   0.12873   0.05846   0.00238
#>  6     5   0.99065   0.12325   0.05846   0.00178
#>  7     6   0.99127   0.11878   0.05726   0.00084
#>  8     7   0.99184   0.11506   0.05492  -0.00049
#>  9     8   0.99238   0.11185   0.05159  -0.00203
#> 10     9   0.99286   0.10929   0.04774  -0.00341
#> # … with 91 more rows
#> 
#> [[56]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.98560   0.15828   0.05930   0.00487
#>  2     1   0.98645   0.15168   0.06253   0.00294
#>  3     2   0.98748   0.14401   0.06431   0.00149
#>  4     3   0.98849   0.13630   0.06559   0.00054
#>  5     4   0.98940   0.12922   0.06632  -0.00025
#>  6     5   0.99018   0.12305   0.06632  -0.00110
#>  7     6   0.99085   0.11813   0.06524  -0.00210
#>  8     7   0.99144   0.11435   0.06290  -0.00322
#>  9     8   0.99197   0.11146   0.05963  -0.00433
#> 10     9   0.99247   0.10905   0.05553  -0.00550
#> # … with 91 more rows
#> 
#> [[57]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99731   0.06738   0.01865  -0.02186
#>  2     1   0.99746   0.06436   0.01994  -0.02323
#>  3     2   0.99760   0.06100   0.02120  -0.02481
#>  4     3   0.99772   0.05801   0.02227  -0.02623
#>  5     4   0.99784   0.05477   0.02350  -0.02769
#>  6     5   0.99793   0.05174   0.02456  -0.02930
#>  7     6   0.99800   0.04888   0.02529  -0.03101
#>  8     7   0.99807   0.04578   0.02602  -0.03277
#>  9     8   0.99812   0.04298   0.02663  -0.03466
#> 10     9   0.99817   0.03983   0.02730  -0.03649
#> # … with 91 more rows
#> 
#> [[58]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99559   0.09262   0.01308   0.00709
#>  2     1   0.99586   0.08931   0.01559   0.00661
#>  3     2   0.99613   0.08592   0.01745   0.00591
#>  4     3   0.99636   0.08301   0.01881   0.00496
#>  5     4   0.99652   0.08081   0.01988   0.00377
#>  6     5   0.99666   0.07899   0.02062   0.00225
#>  7     6   0.99679   0.07729   0.02102   0.00036
#>  8     7   0.99691   0.07559   0.02112  -0.00185
#>  9     8   0.99703   0.07404   0.02082  -0.00432
#> 10     9   0.99715   0.07238   0.02028  -0.00698
#> # … with 91 more rows
#> 
#> [[59]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99395   0.08747   0.06576   0.00951
#>  2     1   0.99444   0.08137   0.06615   0.00957
#>  3     2   0.99486   0.07613   0.06603   0.00953
#>  4     3   0.99523   0.07228   0.06493   0.00924
#>  5     4   0.99552   0.06980   0.06328   0.00856
#>  6     5   0.99577   0.06833   0.06104   0.00732
#>  7     6   0.99603   0.06734   0.05792   0.00565
#>  8     7   0.99631   0.06655   0.05405   0.00359
#>  9     8   0.99659   0.06593   0.04967   0.00128
#> 10     9   0.99682   0.06564   0.04511  -0.00103
#> # … with 91 more rows
#> 
#> [[60]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99425   0.08389   0.06599   0.00853
#>  2     1   0.99452   0.08106   0.06553   0.00752
#>  3     2   0.99480   0.07823   0.06479   0.00680
#>  4     3   0.99506   0.07625   0.06329   0.00573
#>  5     4   0.99529   0.07502   0.06125   0.00434
#>  6     5   0.99552   0.07424   0.05849   0.00267
#>  7     6   0.99574   0.07375   0.05531   0.00084
#>  8     7   0.99597   0.07343   0.05146  -0.00121
#>  9     8   0.99621   0.07306   0.04699  -0.00327
#> 10     9   0.99645   0.07264   0.04217  -0.00519
#> # … with 91 more rows
#> 
#> [[61]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99441   0.09920   0.03375   0.01289
#>  2     1   0.99458   0.09633   0.03718   0.01209
#>  3     2   0.99482   0.09310   0.03933   0.01104
#>  4     3   0.99508   0.08981   0.04070   0.00956
#>  5     4   0.99533   0.08673   0.04164   0.00768
#>  6     5   0.99555   0.08392   0.04247   0.00546
#>  7     6   0.99573   0.08148   0.04326   0.00294
#>  8     7   0.99589   0.07917   0.04397   0.00006
#>  9     8   0.99605   0.07681   0.04454  -0.00305
#> 10     9   0.99620   0.07431   0.04502  -0.00643
#> # … with 91 more rows
#> 
#> [[62]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99448   0.09803   0.03650   0.00843
#>  2     1   0.99475   0.09383   0.03990   0.00820
#>  3     2   0.99499   0.09006   0.04275   0.00741
#>  4     3   0.99518   0.08703   0.04472   0.00614
#>  5     4   0.99535   0.08448   0.04610   0.00435
#>  6     5   0.99551   0.08213   0.04698   0.00214
#>  7     6   0.99568   0.07975   0.04758  -0.00037
#>  8     7   0.99585   0.07726   0.04791  -0.00314
#>  9     8   0.99604   0.07462   0.04801  -0.00608
#> 10     9   0.99623   0.07175   0.04792  -0.00913
#> # … with 91 more rows
#> 
#> [[63]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99368   0.09446   0.06021   0.00780
#>  2     1   0.99426   0.08810   0.06020   0.00794
#>  3     2   0.99471   0.08315   0.05978   0.00746
#>  4     3   0.99514   0.07903   0.05829   0.00661
#>  5     4   0.99553   0.07578   0.05608   0.00538
#>  6     5   0.99587   0.07327   0.05342   0.00378
#>  7     6   0.99619   0.07126   0.05026   0.00181
#>  8     7   0.99648   0.06958   0.04679  -0.00032
#>  9     8   0.99676   0.06801   0.04297  -0.00255
#> 10     9   0.99702   0.06642   0.03882  -0.00483
#> # … with 91 more rows
#> 
#> [[64]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99277   0.10283   0.06109   0.01022
#>  2     1   0.99328   0.09702   0.06216   0.01072
#>  3     2   0.99379   0.09191   0.06186   0.01042
#>  4     3   0.99427   0.08780   0.06022   0.00956
#>  5     4   0.99469   0.08469   0.05786   0.00816
#>  6     5   0.99505   0.08247   0.05508   0.00619
#>  7     6   0.99538   0.08072   0.05185   0.00375
#>  8     7   0.99568   0.07934   0.04823   0.00102
#>  9     8   0.99594   0.07826   0.04440  -0.00173
#> 10     9   0.99617   0.07739   0.04036  -0.00439
#> # … with 91 more rows
#> 
#> attr(,"class")
#> [1] "qts_sample" "list"