QTS Reorientation

This function reorients the quaternions in a QTS for representing attitude with respect to the orientation of the sensor at the first time point.

Usage

reorient_qts(x, disable_normalization = FALSE)

Arguments

x

An object of class qts.

disable_normalization

A boolean specifying whether quaternion normalization should be disabled. Defaults to FALSE.

Value

An object of class qts in which quaternions measure attitude with respect to the orientation of the sensor at the first time point.

Examples

reorient_qts(vespa64$igp[[1]])
#> # A tibble: 101 × 5
#>     time     w                      x                        y         z
#>    <int> <dec>                  <dec>                    <dec>     <dec>
#>  1     0 1     -0.0000000000000000139  0.000000000000000000217  0       
#>  2     1 1.00  -0.00519               -0.00222                 -0.000398
#>  3     2 1.00  -0.0105                -0.00527                 -0.000968
#>  4     3 1.00  -0.0158                -0.00895                 -0.00190 
#>  5     4 1.00  -0.0203                -0.0125                  -0.00307 
#>  6     5 1.00  -0.0240                -0.0159                  -0.00444 
#>  7     6 0.999 -0.0269                -0.0193                  -0.00599 
#>  8     7 0.999 -0.0290                -0.0226                  -0.00777 
#>  9     8 0.999 -0.0307                -0.0260                  -0.00979 
#> 10     9 0.999 -0.0321                -0.0298                  -0.0120  
#> # … with 91 more rows