QTS Random Sampling

This function adds uncorrelated Gaussian noise to the logarithm QTS using an exponential covariance function.

Usage

rnorm_qts(n, mean_qts, alpha = 0.01, beta = 0.001)

Arguments

n

An integer specifying how many QTS should be generated.

mean_qts

An object of class qts specifying the mean QTS.

alpha

A positive scalar specifying the variance of each component of the log-QTS. Defaults to 0.01.

beta

A positive scalar specifying the exponential weight. Defaults to 0.001.

Value

A list of n objects of class qts with added noise as specified by parameters alpha and beta.

Details

See exp_cov_function for details about the roles of alpha and beta in the definition of the covariance operator.

Examples

rnorm_qts(1, vespa64$igp[[1]])
#> [[1]]
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.94948   0.12891   0.27792   0.06805
#>  2     1   0.95119   0.11564   0.27752   0.06965
#>  3     2   0.95546   0.10827   0.26726   0.06286
#>  4     3   0.95658   0.10276   0.26463   0.06614
#>  5     4   0.95668   0.10068   0.26571   0.06347
#>  6     5   0.95608   0.09286   0.27107   0.06171
#>  7     6   0.95665   0.09197   0.26969   0.06026
#>  8     7   0.95603   0.09135   0.27240   0.05889
#>  9     8   0.95751   0.08389   0.27093   0.05232
#> 10     9   0.95810   0.08561   0.26768   0.05534
#> # ℹ 91 more rows
#> 
#> attr(,"class")
#> [1] "qts_sample" "list"