QTS Smoothing via SLERP Interpolation

This function performs a smoothing of a QTS by SLERP interpolation.

Usage

smooth_qts(x, alpha = 0.5)

Arguments

x

An object of class qts.

alpha

A numeric value in [0,1] specifying the amount of smoothing. The closer to one, the smoother the resulting QTS. Defaults to 0.5.

Value

An object of class qts which is a smooth version of the input QTS.

Examples

smooth_qts(vespa64$igp[[1]])
#> # A tibble: 101 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     0   0.99467   0.07637   0.06803   0.01301
#>  2     1   0.99505   0.07302   0.06618   0.01268
#>  3     2   0.99552   0.06888   0.06361   0.01213
#>  4     3   0.99601   0.06453   0.06055   0.01130
#>  5     4   0.99647   0.06045   0.05732   0.01020
#>  6     5   0.99688   0.05684   0.05403   0.00884
#>  7     6   0.99724   0.05380   0.05075   0.00724
#>  8     7   0.99754   0.05128   0.04743   0.00539
#>  9     8   0.99782   0.04912   0.04403   0.00333
#> 10     9   0.99807   0.04709   0.04049   0.00111
#> # … with 91 more rows