The VESPA dataset

A set of QTS representing individual gait patterns (IGPs) of individuals collected under a number of varying factors.

Usage

vespa

Format

A tibble with 320 rows and 7 columns:

V: a categorical variable with two levels specifying the ID of the Volunteer;

E: a categorical variable with two levels specifying the ID of the Experimenter;

S: a categorical variable with four levels specifying the type of Sensor;

P: a categorical variable with four levels specifying the Position of the sensor;

A: a categorical variable with two levels specifying the ID of the Acquisition pathway;

R: a categorical variable with 5 levels specifying the ID of the Repetition;

igp: A 101x5 tibble storing a QTS which represents the IGP of the individual under a specific set of VESPA conditions.

Details

The IGP measures the hip rotation during a typical gait cycle. Each rotation is expressed with respect to the mean position of the sensor during the gait cycle. Each IGP is then straightened so that it is periodic with a last point matching the first one.