QTS Differentiation

This function computes the first derivative of quaternion time series with respect to time.

Usage

differentiate(x)

# S3 method for class 'qts'
differentiate(x)

# S3 method for class 'qts_sample'
differentiate(x)

Arguments

x

An object of class qts or qts_sample.

Value

An object of the same class as the input argument x in which quaternions measure the rotation to be applied to transform attitude at previous time point to attitude at current time point.

Examples

differentiate(vespa64$igp[[1]])
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99998  -0.00519  -0.00222  -0.00040
#>  2     2   0.99998  -0.00529  -0.00305  -0.00057
#>  3     3   0.99998  -0.00534  -0.00367  -0.00094
#>  4     4   0.99998  -0.00447  -0.00352  -0.00119
#>  5     5   0.99999  -0.00374  -0.00345  -0.00140
#>  6     6   0.99999  -0.00292  -0.00331  -0.00158
#>  7     7   0.99999  -0.00212  -0.00327  -0.00182
#>  8     8   0.99999  -0.00166  -0.00341  -0.00209
#>  9     9   0.99999  -0.00145  -0.00371  -0.00229
#> 10    10   0.99999  -0.00168  -0.00394  -0.00239
#> # ℹ 90 more rows
differentiate(vespa64$igp)
#> [[1]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99998  -0.00519  -0.00222  -0.00040
#>  2     2   0.99998  -0.00529  -0.00305  -0.00057
#>  3     3   0.99998  -0.00534  -0.00367  -0.00094
#>  4     4   0.99998  -0.00447  -0.00352  -0.00119
#>  5     5   0.99999  -0.00374  -0.00345  -0.00140
#>  6     6   0.99999  -0.00292  -0.00331  -0.00158
#>  7     7   0.99999  -0.00212  -0.00327  -0.00182
#>  8     8   0.99999  -0.00166  -0.00341  -0.00209
#>  9     9   0.99999  -0.00145  -0.00371  -0.00229
#> 10    10   0.99999  -0.00168  -0.00394  -0.00239
#> # ℹ 90 more rows
#> 
#> [[2]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99998  -0.00562  -0.00244   0.00021
#>  2     2   0.99998  -0.00562  -0.00285   0.00006
#>  3     3   0.99998  -0.00577  -0.00337  -0.00028
#>  4     4   0.99998  -0.00569  -0.00364  -0.00055
#>  5     5   0.99998  -0.00508  -0.00369  -0.00088
#>  6     6   0.99999  -0.00380  -0.00343  -0.00120
#>  7     7   0.99999  -0.00288  -0.00344  -0.00154
#>  8     8   0.99999  -0.00221  -0.00365  -0.00188
#>  9     9   0.99999  -0.00171  -0.00389  -0.00215
#> 10    10   0.99999  -0.00133  -0.00390  -0.00228
#> # ℹ 90 more rows
#> 
#> [[3]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99998  -0.00564   0.00059  -0.00011
#>  2     2   0.99998  -0.00653  -0.00025   0.00072
#>  3     3   0.99997  -0.00744  -0.00093   0.00096
#>  4     4   0.99997  -0.00702  -0.00130   0.00112
#>  5     5   0.99998  -0.00613  -0.00167   0.00103
#>  6     6   0.99998  -0.00524  -0.00202   0.00100
#>  7     7   0.99999  -0.00422  -0.00243   0.00097
#>  8     8   0.99999  -0.00337  -0.00290   0.00093
#>  9     9   0.99999  -0.00273  -0.00343   0.00088
#> 10    10   0.99999  -0.00217  -0.00386   0.00082
#> # ℹ 90 more rows
#> 
#> [[4]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99998  -0.00579   0.00081   0.00039
#>  2     2   0.99998  -0.00622  -0.00025   0.00075
#>  3     3   0.99998  -0.00672  -0.00079   0.00097
#>  4     4   0.99998  -0.00668  -0.00127   0.00110
#>  5     5   0.99998  -0.00595  -0.00164   0.00107
#>  6     6   0.99999  -0.00497  -0.00197   0.00101
#>  7     7   0.99999  -0.00384  -0.00233   0.00092
#>  8     8   0.99999  -0.00290  -0.00278   0.00089
#>  9     9   0.99999  -0.00212  -0.00346   0.00084
#> 10    10   0.99999  -0.00164  -0.00408   0.00069
#> # ℹ 90 more rows
#> 
#> [[5]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99999  -0.00494  -0.00206  -0.00098
#>  2     2   0.99999  -0.00476  -0.00252  -0.00043
#>  3     3   0.99998  -0.00460  -0.00319  -0.00053
#>  4     4   0.99998  -0.00427  -0.00347  -0.00086
#>  5     5   0.99999  -0.00392  -0.00362  -0.00121
#>  6     6   0.99999  -0.00341  -0.00363  -0.00155
#>  7     7   0.99999  -0.00265  -0.00348  -0.00185
#>  8     8   0.99999  -0.00195  -0.00346  -0.00213
#>  9     9   0.99999  -0.00150  -0.00362  -0.00232
#> 10    10   0.99999  -0.00147  -0.00385  -0.00245
#> # ℹ 90 more rows
#> 
#> [[6]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99998  -0.00559  -0.00017  -0.00074
#>  2     2   0.99998  -0.00554  -0.00178  -0.00025
#>  3     3   0.99998  -0.00576  -0.00300  -0.00024
#>  4     4   0.99998  -0.00560  -0.00360  -0.00046
#>  5     5   0.99998  -0.00499  -0.00363  -0.00068
#>  6     6   0.99998  -0.00423  -0.00354  -0.00099
#>  7     7   0.99999  -0.00294  -0.00330  -0.00145
#>  8     8   0.99999  -0.00196  -0.00327  -0.00184
#>  9     9   0.99999  -0.00145  -0.00353  -0.00224
#> 10    10   0.99999  -0.00113  -0.00376  -0.00254
#> # ℹ 90 more rows
#> 
#> [[7]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99999  -0.00372   0.00203  -0.00002
#>  2     2   0.99999  -0.00331   0.00219   0.00036
#>  3     3   0.99999  -0.00327   0.00189   0.00090
#>  4     4   0.99999  -0.00338   0.00139   0.00124
#>  5     5   0.99999  -0.00332   0.00087   0.00151
#>  6     6   0.99999  -0.00322   0.00051   0.00189
#>  7     7   0.99999  -0.00281  -0.00032   0.00198
#>  8     8   0.99999  -0.00231  -0.00112   0.00200
#>  9     9   0.99999  -0.00204  -0.00167   0.00197
#> 10    10   0.99999  -0.00195  -0.00211   0.00199
#> # ℹ 90 more rows
#> 
#> [[8]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99999  -0.00268   0.00201   0.00192
#>  2     2   0.99999  -0.00331   0.00215   0.00235
#>  3     3   0.99999  -0.00350   0.00154   0.00239
#>  4     4   0.99999  -0.00351   0.00084   0.00235
#>  5     5   0.99999  -0.00331   0.00001   0.00229
#>  6     6   0.99999  -0.00281  -0.00100   0.00218
#>  7     7   0.99999  -0.00231  -0.00198   0.00208
#>  8     8   0.99999  -0.00176  -0.00286   0.00203
#>  9     9   0.99999  -0.00141  -0.00365   0.00199
#> 10    10   0.99999  -0.00131  -0.00417   0.00196
#> # ℹ 90 more rows
#> 
#> [[9]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99998  -0.00660  -0.00025   0.00077
#>  2     2   0.99997  -0.00756  -0.00152   0.00041
#>  3     3   0.99996  -0.00843  -0.00282   0.00005
#>  4     4   0.99996  -0.00806  -0.00360  -0.00054
#>  5     5   0.99997  -0.00708  -0.00399  -0.00117
#>  6     6   0.99998  -0.00532  -0.00402  -0.00195
#>  7     7   0.99998  -0.00340  -0.00364  -0.00268
#>  8     8   0.99999  -0.00181  -0.00309  -0.00341
#>  9     9   0.99999  -0.00083  -0.00239  -0.00385
#> 10    10   0.99999  -0.00050  -0.00168  -0.00391
#> # ℹ 90 more rows
#> 
#> [[10]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99997  -0.00770   0.00050   0.00029
#>  2     2   0.99997  -0.00795  -0.00116   0.00050
#>  3     3   0.99996  -0.00800  -0.00262   0.00012
#>  4     4   0.99996  -0.00788  -0.00348  -0.00040
#>  5     5   0.99997  -0.00702  -0.00386  -0.00105
#>  6     6   0.99998  -0.00560  -0.00387  -0.00179
#>  7     7   0.99998  -0.00370  -0.00348  -0.00242
#>  8     8   0.99999  -0.00199  -0.00300  -0.00292
#>  9     9   0.99999  -0.00051  -0.00242  -0.00337
#> 10    10   0.99999   0.00011  -0.00186  -0.00363
#> # ℹ 90 more rows
#> 
#> [[11]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99997  -0.00743   0.00259  -0.00012
#>  2     2   0.99997  -0.00724   0.00165   0.00007
#>  3     3   0.99998  -0.00657   0.00033   0.00007
#>  4     4   0.99999  -0.00510  -0.00094  -0.00007
#>  5     5   0.99999  -0.00389  -0.00223  -0.00030
#>  6     6   0.99999  -0.00302  -0.00371  -0.00060
#>  7     7   0.99998  -0.00288  -0.00490  -0.00084
#>  8     8   0.99998  -0.00271  -0.00547  -0.00112
#>  9     9   0.99998  -0.00243  -0.00535  -0.00144
#> 10    10   0.99999  -0.00190  -0.00428  -0.00155
#> # ℹ 90 more rows
#> 
#> [[12]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99996  -0.00766   0.00346  -0.00017
#>  2     2   0.99997  -0.00745   0.00226   0.00036
#>  3     3   0.99998  -0.00668   0.00066   0.00054
#>  4     4   0.99998  -0.00573  -0.00129   0.00062
#>  5     5   0.99998  -0.00520  -0.00306   0.00057
#>  6     6   0.99998  -0.00460  -0.00444   0.00026
#>  7     7   0.99998  -0.00401  -0.00523  -0.00014
#>  8     8   0.99998  -0.00307  -0.00532  -0.00073
#>  9     9   0.99999  -0.00213  -0.00438  -0.00133
#> 10    10   0.99999  -0.00160  -0.00352  -0.00178
#> # ℹ 90 more rows
#> 
#> [[13]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99999  -0.00362  -0.00259   0.00018
#>  2     2   0.99998  -0.00401  -0.00414   0.00031
#>  3     3   0.99999  -0.00338  -0.00369  -0.00041
#>  4     4   0.99999  -0.00253  -0.00303  -0.00109
#>  5     5   0.99999  -0.00155  -0.00226  -0.00181
#>  6     6   1.00000  -0.00081  -0.00163  -0.00231
#>  7     7   1.00000  -0.00041  -0.00123  -0.00268
#>  8     8   0.99999  -0.00043  -0.00110  -0.00300
#>  9     9   0.99999  -0.00087  -0.00117  -0.00314
#> 10    10   0.99999  -0.00146  -0.00136  -0.00300
#> # ℹ 90 more rows
#> 
#> [[14]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99999  -0.00359  -0.00178   0.00074
#>  2     2   0.99999  -0.00418  -0.00255   0.00033
#>  3     3   0.99999  -0.00384  -0.00318  -0.00021
#>  4     4   0.99999  -0.00315  -0.00334  -0.00080
#>  5     5   0.99999  -0.00225  -0.00319  -0.00142
#>  6     6   0.99999  -0.00133  -0.00294  -0.00198
#>  7     7   0.99999  -0.00067  -0.00269  -0.00234
#>  8     8   0.99999  -0.00046  -0.00252  -0.00257
#>  9     9   0.99999  -0.00071  -0.00245  -0.00270
#> 10    10   0.99999  -0.00134  -0.00235  -0.00271
#> # ℹ 90 more rows
#> 
#> [[15]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99999  -0.00455  -0.00186   0.00107
#>  2     2   0.99999  -0.00417  -0.00301   0.00094
#>  3     3   0.99999  -0.00350  -0.00384   0.00040
#>  4     4   0.99999  -0.00267  -0.00426  -0.00022
#>  5     5   0.99999  -0.00167  -0.00424  -0.00088
#>  6     6   0.99999  -0.00074  -0.00403  -0.00150
#>  7     7   0.99999  -0.00014  -0.00371  -0.00194
#>  8     8   0.99999   0.00010  -0.00333  -0.00217
#>  9     9   0.99999  -0.00011  -0.00291  -0.00215
#> 10    10   0.99999  -0.00070  -0.00251  -0.00198
#> # ℹ 90 more rows
#> 
#> [[16]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99998  -0.00546  -0.00077   0.00127
#>  2     2   0.99999  -0.00472  -0.00212   0.00115
#>  3     3   0.99999  -0.00414  -0.00309   0.00074
#>  4     4   0.99999  -0.00333  -0.00364   0.00027
#>  5     5   0.99999  -0.00238  -0.00387  -0.00028
#>  6     6   0.99999  -0.00138  -0.00380  -0.00085
#>  7     7   0.99999  -0.00069  -0.00371  -0.00139
#>  8     8   0.99999  -0.00038  -0.00365  -0.00181
#>  9     9   0.99999  -0.00051  -0.00369  -0.00202
#> 10    10   0.99999  -0.00096  -0.00373  -0.00206
#> # ℹ 90 more rows
#> 
#> [[17]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99999  -0.00430   0.00027  -0.00133
#>  2     2   0.99999  -0.00398  -0.00032  -0.00147
#>  3     3   0.99999  -0.00398  -0.00032  -0.00173
#>  4     4   0.99999  -0.00376  -0.00028  -0.00193
#>  5     5   0.99999  -0.00350  -0.00015  -0.00225
#>  6     6   0.99999  -0.00310  -0.00002  -0.00255
#>  7     7   0.99999  -0.00276   0.00003  -0.00282
#>  8     8   0.99999  -0.00259   0.00007  -0.00299
#>  9     9   0.99999  -0.00253   0.00007  -0.00309
#> 10    10   0.99999  -0.00255   0.00006  -0.00311
#> # ℹ 90 more rows
#> 
#> [[18]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99999  -0.00479  -0.00072  -0.00115
#>  2     2   0.99999  -0.00443  -0.00084  -0.00139
#>  3     3   0.99999  -0.00412  -0.00062  -0.00177
#>  4     4   0.99999  -0.00368  -0.00041  -0.00205
#>  5     5   0.99999  -0.00323  -0.00032  -0.00242
#>  6     6   0.99999  -0.00271  -0.00027  -0.00276
#>  7     7   0.99999  -0.00223  -0.00016  -0.00303
#>  8     8   0.99999  -0.00218  -0.00007  -0.00333
#>  9     9   0.99999  -0.00217  -0.00010  -0.00337
#> 10    10   0.99999  -0.00231  -0.00017  -0.00337
#> # ℹ 90 more rows
#> 
#> [[19]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99999  -0.00375   0.00135   0.00012
#>  2     2   0.99999  -0.00387   0.00020   0.00040
#>  3     3   0.99999  -0.00440  -0.00105   0.00059
#>  4     4   0.99999  -0.00465  -0.00186   0.00058
#>  5     5   0.99999  -0.00476  -0.00249   0.00047
#>  6     6   0.99999  -0.00459  -0.00294   0.00021
#>  7     7   0.99998  -0.00443  -0.00338  -0.00021
#>  8     8   0.99998  -0.00408  -0.00378  -0.00070
#>  9     9   0.99998  -0.00366  -0.00398  -0.00108
#> 10    10   0.99998  -0.00341  -0.00410  -0.00141
#> # ℹ 90 more rows
#> 
#> [[20]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99998  -0.00539  -0.00145   0.00035
#>  2     2   0.99998  -0.00518  -0.00239   0.00063
#>  3     3   0.99998  -0.00525  -0.00303   0.00032
#>  4     4   0.99998  -0.00481  -0.00327  -0.00015
#>  5     5   0.99999  -0.00399  -0.00334  -0.00063
#>  6     6   0.99999  -0.00330  -0.00348  -0.00104
#>  7     7   0.99999  -0.00296  -0.00382  -0.00140
#>  8     8   0.99999  -0.00269  -0.00426  -0.00174
#>  9     9   0.99998  -0.00253  -0.00448  -0.00192
#> 10    10   0.99998  -0.00262  -0.00445  -0.00193
#> # ℹ 90 more rows
#> 
#> [[21]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99998  -0.00559  -0.00068  -0.00068
#>  2     2   0.99999  -0.00501  -0.00088  -0.00076
#>  3     3   0.99999  -0.00484  -0.00074  -0.00117
#>  4     4   0.99999  -0.00436  -0.00057  -0.00157
#>  5     5   0.99999  -0.00400  -0.00036  -0.00210
#>  6     6   0.99999  -0.00370  -0.00021  -0.00262
#>  7     7   0.99999  -0.00351  -0.00014  -0.00294
#>  8     8   0.99999  -0.00331  -0.00006  -0.00307
#>  9     9   0.99999  -0.00333  -0.00002  -0.00317
#> 10    10   0.99999  -0.00357  -0.00001  -0.00326
#> # ℹ 90 more rows
#> 
#> [[22]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99999  -0.00488  -0.00124  -0.00039
#>  2     2   0.99999  -0.00516  -0.00137  -0.00081
#>  3     3   0.99999  -0.00502  -0.00103  -0.00128
#>  4     4   0.99999  -0.00435  -0.00052  -0.00173
#>  5     5   0.99999  -0.00371  -0.00026  -0.00216
#>  6     6   0.99999  -0.00320  -0.00026  -0.00262
#>  7     7   0.99999  -0.00267  -0.00023  -0.00284
#>  8     8   0.99999  -0.00250  -0.00022  -0.00299
#>  9     9   0.99999  -0.00278  -0.00030  -0.00314
#> 10    10   0.99999  -0.00323  -0.00042  -0.00321
#> # ℹ 90 more rows
#> 
#> [[23]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99999  -0.00424   0.00085   0.00012
#>  2     2   0.99999  -0.00405  -0.00038   0.00075
#>  3     3   0.99999  -0.00438  -0.00145   0.00091
#>  4     4   0.99999  -0.00468  -0.00234   0.00078
#>  5     5   0.99998  -0.00468  -0.00292   0.00043
#>  6     6   0.99998  -0.00455  -0.00336  -0.00016
#>  7     7   0.99998  -0.00425  -0.00358  -0.00065
#>  8     8   0.99998  -0.00397  -0.00391  -0.00116
#>  9     9   0.99998  -0.00370  -0.00425  -0.00169
#> 10    10   0.99998  -0.00372  -0.00450  -0.00211
#> # ℹ 90 more rows
#> 
#> [[24]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99999  -0.00460   0.00025  -0.00041
#>  2     2   0.99999  -0.00394  -0.00033   0.00045
#>  3     3   0.99999  -0.00415  -0.00139   0.00054
#>  4     4   0.99999  -0.00448  -0.00233   0.00056
#>  5     5   0.99998  -0.00461  -0.00307   0.00036
#>  6     6   0.99998  -0.00457  -0.00363  -0.00001
#>  7     7   0.99998  -0.00422  -0.00396  -0.00051
#>  8     8   0.99998  -0.00391  -0.00431  -0.00101
#>  9     9   0.99998  -0.00363  -0.00463  -0.00147
#> 10    10   0.99998  -0.00336  -0.00476  -0.00182
#> # ℹ 90 more rows
#> 
#> [[25]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   1.00000  -0.00076  -0.00245  -0.00075
#>  2     2   0.99999  -0.00216  -0.00358  -0.00064
#>  3     3   0.99999  -0.00303  -0.00399  -0.00058
#>  4     4   0.99998  -0.00384  -0.00431  -0.00056
#>  5     5   0.99998  -0.00477  -0.00462  -0.00057
#>  6     6   0.99997  -0.00592  -0.00492  -0.00047
#>  7     7   0.99997  -0.00659  -0.00483  -0.00050
#>  8     8   0.99996  -0.00708  -0.00479  -0.00050
#>  9     9   0.99996  -0.00740  -0.00468  -0.00051
#> 10    10   0.99996  -0.00733  -0.00422  -0.00055
#> # ℹ 90 more rows
#> 
#> [[26]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99999  -0.00336  -0.00371   0.00095
#>  2     2   0.99999  -0.00349  -0.00373   0.00054
#>  3     3   0.99999  -0.00351  -0.00351   0.00027
#>  4     4   0.99999  -0.00375  -0.00327  -0.00004
#>  5     5   0.99999  -0.00314  -0.00286  -0.00027
#>  6     6   0.99999  -0.00238  -0.00248  -0.00059
#>  7     7   0.99999  -0.00201  -0.00250  -0.00083
#>  8     8   0.99999  -0.00167  -0.00273  -0.00105
#>  9     9   0.99999  -0.00157  -0.00291  -0.00121
#> 10    10   0.99999  -0.00205  -0.00331  -0.00134
#> # ℹ 90 more rows
#> 
#> [[27]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99999  -0.00138  -0.00301   0.00256
#>  2     2   0.99999  -0.00063  -0.00353   0.00230
#>  3     3   0.99999   0.00020  -0.00346   0.00162
#>  4     4   0.99999   0.00080  -0.00346   0.00123
#>  5     5   0.99999   0.00133  -0.00332   0.00077
#>  6     6   0.99999   0.00139  -0.00302   0.00025
#>  7     7   0.99999   0.00109  -0.00302  -0.00008
#>  8     8   0.99999   0.00072  -0.00320  -0.00024
#>  9     9   0.99999   0.00024  -0.00346  -0.00024
#> 10    10   0.99999  -0.00017  -0.00362  -0.00005
#> # ℹ 90 more rows
#> 
#> [[28]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99999  -0.00342  -0.00167   0.00187
#>  2     2   0.99999  -0.00212  -0.00208   0.00220
#>  3     3   0.99999  -0.00128  -0.00234   0.00218
#>  4     4   1.00000  -0.00035  -0.00227   0.00193
#>  5     5   1.00000   0.00054  -0.00221   0.00145
#>  6     6   1.00000   0.00112  -0.00239   0.00106
#>  7     7   1.00000   0.00141  -0.00258   0.00055
#>  8     8   0.99999   0.00120  -0.00294   0.00004
#>  9     9   0.99999   0.00081  -0.00334  -0.00027
#> 10    10   0.99999   0.00017  -0.00370  -0.00039
#> # ℹ 90 more rows
#> 
#> [[29]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99999  -0.00188  -0.00288   0.00262
#>  2     2   0.99999  -0.00207  -0.00346   0.00262
#>  3     3   0.99999  -0.00142  -0.00329   0.00218
#>  4     4   0.99999  -0.00070  -0.00272   0.00155
#>  5     5   1.00000   0.00001  -0.00206   0.00077
#>  6     6   1.00000   0.00044  -0.00187   0.00017
#>  7     7   1.00000   0.00061  -0.00194  -0.00030
#>  8     8   1.00000   0.00052  -0.00236  -0.00060
#>  9     9   1.00000   0.00022  -0.00294  -0.00076
#> 10    10   0.99999  -0.00026  -0.00359  -0.00079
#> # ℹ 90 more rows
#> 
#> [[30]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99999  -0.00021  -0.00267   0.00205
#>  2     2   0.99999  -0.00152  -0.00249   0.00191
#>  3     3   0.99999  -0.00097  -0.00261   0.00150
#>  4     4   1.00000  -0.00030  -0.00240   0.00098
#>  5     5   1.00000   0.00027  -0.00234   0.00046
#>  6     6   1.00000   0.00066  -0.00232  -0.00005
#>  7     7   1.00000   0.00076  -0.00255  -0.00036
#>  8     8   1.00000   0.00059  -0.00298  -0.00058
#>  9     9   0.99999   0.00022  -0.00357  -0.00061
#> 10    10   0.99999  -0.00027  -0.00408  -0.00057
#> # ℹ 90 more rows
#> 
#> [[31]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99999  -0.00156  -0.00087   0.00297
#>  2     2   1.00000  -0.00088  -0.00117   0.00272
#>  3     3   1.00000  -0.00039  -0.00158   0.00232
#>  4     4   1.00000   0.00024  -0.00185   0.00176
#>  5     5   1.00000   0.00071  -0.00203   0.00129
#>  6     6   1.00000   0.00089  -0.00235   0.00106
#>  7     7   1.00000   0.00089  -0.00273   0.00091
#>  8     8   0.99999   0.00064  -0.00328   0.00086
#>  9     9   0.99999   0.00031  -0.00380   0.00098
#> 10    10   0.99999  -0.00005  -0.00422   0.00113
#> # ℹ 90 more rows
#> 
#> [[32]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   1.00000  -0.00144  -0.00175   0.00200
#>  2     2   1.00000  -0.00037  -0.00163   0.00194
#>  3     3   1.00000   0.00015  -0.00166   0.00162
#>  4     4   1.00000   0.00078  -0.00183   0.00118
#>  5     5   1.00000   0.00120  -0.00223   0.00087
#>  6     6   0.99999   0.00143  -0.00274   0.00071
#>  7     7   0.99999   0.00147  -0.00332   0.00069
#>  8     8   0.99999   0.00131  -0.00393   0.00075
#>  9     9   0.99999   0.00087  -0.00443   0.00093
#> 10    10   0.99999   0.00037  -0.00482   0.00115
#> # ℹ 90 more rows
#> 
#> [[33]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99998  -0.00488   0.00358   0.00011
#>  2     2   0.99998  -0.00468   0.00374  -0.00010
#>  3     3   0.99998  -0.00455   0.00343  -0.00052
#>  4     4   0.99999  -0.00412   0.00317  -0.00088
#>  5     5   0.99999  -0.00357   0.00300  -0.00125
#>  6     6   0.99999  -0.00312   0.00276  -0.00171
#>  7     7   0.99999  -0.00291   0.00247  -0.00219
#>  8     8   0.99999  -0.00289   0.00216  -0.00266
#>  9     9   0.99999  -0.00293   0.00184  -0.00310
#> 10    10   0.99999  -0.00340   0.00139  -0.00348
#> # ℹ 90 more rows
#> 
#> [[34]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99999  -0.00238   0.00323   0.00005
#>  2     2   0.99999  -0.00310   0.00300  -0.00072
#>  3     3   0.99999  -0.00285   0.00287  -0.00125
#>  4     4   0.99999  -0.00234   0.00254  -0.00178
#>  5     5   0.99999  -0.00220   0.00227  -0.00225
#>  6     6   0.99999  -0.00233   0.00204  -0.00274
#>  7     7   0.99999  -0.00259   0.00155  -0.00315
#>  8     8   0.99999  -0.00289   0.00108  -0.00335
#>  9     9   0.99999  -0.00313   0.00073  -0.00354
#> 10    10   0.99999  -0.00336   0.00050  -0.00369
#> # ℹ 90 more rows
#> 
#> [[35]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99999  -0.00443   0.00175  -0.00145
#>  2     2   0.99999  -0.00436   0.00112  -0.00199
#>  3     3   0.99999  -0.00411   0.00044  -0.00215
#>  4     4   0.99999  -0.00351  -0.00014  -0.00230
#>  5     5   0.99999  -0.00268  -0.00083  -0.00239
#>  6     6   0.99999  -0.00240  -0.00150  -0.00263
#>  7     7   0.99999  -0.00209  -0.00207  -0.00267
#>  8     8   0.99999  -0.00211  -0.00282  -0.00287
#>  9     9   0.99999  -0.00230  -0.00366  -0.00313
#> 10    10   0.99998  -0.00248  -0.00407  -0.00331
#> # ℹ 90 more rows
#> 
#> [[36]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99999  -0.00473   0.00140  -0.00237
#>  2     2   0.99999  -0.00476   0.00092  -0.00247
#>  3     3   0.99999  -0.00429  -0.00009  -0.00245
#>  4     4   0.99999  -0.00364  -0.00103  -0.00239
#>  5     5   0.99999  -0.00314  -0.00206  -0.00247
#>  6     6   0.99999  -0.00293  -0.00301  -0.00266
#>  7     7   0.99998  -0.00271  -0.00384  -0.00291
#>  8     8   0.99998  -0.00264  -0.00443  -0.00305
#>  9     9   0.99998  -0.00262  -0.00491  -0.00312
#> 10    10   0.99998  -0.00272  -0.00509  -0.00315
#> # ℹ 90 more rows
#> 
#> [[37]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99998  -0.00405   0.00431   0.00153
#>  2     2   0.99998  -0.00447   0.00433   0.00136
#>  3     3   0.99998  -0.00415   0.00429   0.00068
#>  4     4   0.99999  -0.00322   0.00387   0.00000
#>  5     5   0.99999  -0.00245   0.00368  -0.00052
#>  6     6   0.99999  -0.00207   0.00348  -0.00108
#>  7     7   0.99999  -0.00201   0.00328  -0.00172
#>  8     8   0.99999  -0.00223   0.00308  -0.00238
#>  9     9   0.99999  -0.00243   0.00271  -0.00286
#> 10    10   0.99999  -0.00264   0.00230  -0.00317
#> # ℹ 90 more rows
#> 
#> [[38]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99998  -0.00425   0.00425   0.00062
#>  2     2   0.99998  -0.00416   0.00369   0.00015
#>  3     3   0.99999  -0.00394   0.00306  -0.00044
#>  4     4   0.99999  -0.00372   0.00287  -0.00096
#>  5     5   0.99999  -0.00371   0.00270  -0.00153
#>  6     6   0.99999  -0.00373   0.00240  -0.00197
#>  7     7   0.99999  -0.00355   0.00207  -0.00235
#>  8     8   0.99999  -0.00357   0.00178  -0.00277
#>  9     9   0.99999  -0.00373   0.00153  -0.00325
#> 10    10   0.99999  -0.00374   0.00126  -0.00361
#> # ℹ 90 more rows
#> 
#> [[39]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99999  -0.00209   0.00312  -0.00162
#>  2     2   0.99999  -0.00339   0.00228  -0.00137
#>  3     3   0.99999  -0.00353   0.00140  -0.00169
#>  4     4   0.99999  -0.00331   0.00046  -0.00196
#>  5     5   0.99999  -0.00303  -0.00058  -0.00227
#>  6     6   0.99999  -0.00286  -0.00125  -0.00254
#>  7     7   0.99999  -0.00292  -0.00190  -0.00288
#>  8     8   0.99999  -0.00309  -0.00249  -0.00322
#>  9     9   0.99999  -0.00311  -0.00285  -0.00328
#> 10    10   0.99998  -0.00312  -0.00323  -0.00332
#> # ℹ 90 more rows
#> 
#> [[40]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99997  -0.00515   0.00471  -0.00186
#>  2     2   0.99998  -0.00468   0.00307  -0.00183
#>  3     3   0.99999  -0.00398   0.00175  -0.00187
#>  4     4   0.99999  -0.00334   0.00052  -0.00196
#>  5     5   0.99999  -0.00284  -0.00070  -0.00216
#>  6     6   0.99999  -0.00253  -0.00191  -0.00241
#>  7     7   0.99999  -0.00237  -0.00294  -0.00265
#>  8     8   0.99999  -0.00249  -0.00388  -0.00294
#>  9     9   0.99998  -0.00267  -0.00457  -0.00312
#> 10    10   0.99998  -0.00287  -0.00492  -0.00316
#> # ℹ 90 more rows
#> 
#> [[41]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   1.00000  -0.00235   0.00084  -0.00171
#>  2     2   0.99999  -0.00275   0.00010  -0.00189
#>  3     3   0.99999  -0.00275  -0.00090  -0.00201
#>  4     4   0.99999  -0.00255  -0.00196  -0.00222
#>  5     5   0.99999  -0.00226  -0.00291  -0.00238
#>  6     6   0.99999  -0.00216  -0.00346  -0.00262
#>  7     7   0.99999  -0.00216  -0.00379  -0.00294
#>  8     8   0.99998  -0.00229  -0.00396  -0.00307
#>  9     9   0.99998  -0.00270  -0.00414  -0.00320
#> 10    10   0.99998  -0.00295  -0.00405  -0.00327
#> # ℹ 90 more rows
#> 
#> [[42]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99998  -0.00533  -0.00027  -0.00156
#>  2     2   0.99999  -0.00432  -0.00113  -0.00164
#>  3     3   0.99999  -0.00338  -0.00160  -0.00179
#>  4     4   0.99999  -0.00257  -0.00191  -0.00194
#>  5     5   0.99999  -0.00206  -0.00221  -0.00210
#>  6     6   0.99999  -0.00193  -0.00252  -0.00230
#>  7     7   0.99999  -0.00189  -0.00275  -0.00239
#>  8     8   0.99999  -0.00216  -0.00312  -0.00239
#>  9     9   0.99999  -0.00283  -0.00360  -0.00251
#> 10    10   0.99998  -0.00391  -0.00394  -0.00259
#> # ℹ 90 more rows
#> 
#> [[43]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99997  -0.00707   0.00004  -0.00247
#>  2     2   0.99998  -0.00663  -0.00045  -0.00211
#>  3     3   0.99998  -0.00584  -0.00098  -0.00173
#>  4     4   0.99999  -0.00474  -0.00147  -0.00148
#>  5     5   0.99999  -0.00382  -0.00199  -0.00131
#>  6     6   0.99999  -0.00309  -0.00243  -0.00117
#>  7     7   0.99999  -0.00247  -0.00293  -0.00104
#>  8     8   0.99999  -0.00225  -0.00364  -0.00095
#>  9     9   0.99999  -0.00230  -0.00437  -0.00086
#> 10    10   0.99998  -0.00256  -0.00503  -0.00057
#> # ℹ 90 more rows
#> 
#> [[44]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99997  -0.00695  -0.00014  -0.00269
#>  2     2   0.99998  -0.00514  -0.00049  -0.00185
#>  3     3   0.99999  -0.00380  -0.00085  -0.00146
#>  4     4   0.99999  -0.00281  -0.00138  -0.00123
#>  5     5   0.99999  -0.00244  -0.00199  -0.00110
#>  6     6   0.99999  -0.00231  -0.00279  -0.00089
#>  7     7   0.99999  -0.00230  -0.00344  -0.00067
#>  8     8   0.99999  -0.00236  -0.00402  -0.00041
#>  9     9   0.99999  -0.00251  -0.00446  -0.00017
#> 10    10   0.99998  -0.00282  -0.00477   0.00009
#> # ℹ 90 more rows
#> 
#> [[45]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99998  -0.00476   0.00215  -0.00171
#>  2     2   0.99999  -0.00484   0.00147  -0.00183
#>  3     3   0.99999  -0.00422   0.00073  -0.00194
#>  4     4   0.99999  -0.00349   0.00004  -0.00205
#>  5     5   0.99999  -0.00272  -0.00051  -0.00227
#>  6     6   0.99999  -0.00215  -0.00094  -0.00241
#>  7     7   0.99999  -0.00190  -0.00131  -0.00251
#>  8     8   0.99999  -0.00170  -0.00173  -0.00255
#>  9     9   0.99999  -0.00163  -0.00214  -0.00255
#> 10    10   0.99999  -0.00176  -0.00260  -0.00262
#> # ℹ 90 more rows
#> 
#> [[46]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99998  -0.00419   0.00323  -0.00229
#>  2     2   0.99998  -0.00512   0.00281  -0.00230
#>  3     3   0.99998  -0.00471   0.00196  -0.00211
#>  4     4   0.99999  -0.00438   0.00102  -0.00211
#>  5     5   0.99999  -0.00367   0.00035  -0.00214
#>  6     6   0.99999  -0.00297  -0.00005  -0.00235
#>  7     7   0.99999  -0.00218  -0.00054  -0.00260
#>  8     8   0.99999  -0.00152  -0.00096  -0.00269
#>  9     9   0.99999  -0.00125  -0.00132  -0.00294
#> 10    10   1.00000  -0.00085  -0.00191  -0.00235
#> # ℹ 90 more rows
#> 
#> [[47]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99998  -0.00573   0.00143  -0.00189
#>  2     2   0.99998  -0.00633   0.00121  -0.00232
#>  3     3   0.99998  -0.00517   0.00008  -0.00220
#>  4     4   0.99999  -0.00397  -0.00064  -0.00219
#>  5     5   0.99999  -0.00285  -0.00117  -0.00217
#>  6     6   0.99999  -0.00227  -0.00165  -0.00216
#>  7     7   0.99999  -0.00198  -0.00224  -0.00213
#>  8     8   0.99999  -0.00166  -0.00286  -0.00197
#>  9     9   0.99999  -0.00151  -0.00330  -0.00176
#> 10    10   0.99999  -0.00154  -0.00365  -0.00150
#> # ℹ 90 more rows
#> 
#> [[48]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99998  -0.00462   0.00241  -0.00275
#>  2     2   0.99999  -0.00435   0.00225  -0.00239
#>  3     3   0.99999  -0.00368   0.00117  -0.00206
#>  4     4   0.99999  -0.00298   0.00016  -0.00180
#>  5     5   1.00000  -0.00231  -0.00055  -0.00159
#>  6     6   1.00000  -0.00180  -0.00110  -0.00140
#>  7     7   1.00000  -0.00114  -0.00171  -0.00126
#>  8     8   1.00000  -0.00066  -0.00213  -0.00117
#>  9     9   1.00000  -0.00055  -0.00244  -0.00111
#> 10    10   1.00000  -0.00065  -0.00288  -0.00109
#> # ℹ 90 more rows
#> 
#> [[49]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99997  -0.00720   0.00242  -0.00153
#>  2     2   0.99997  -0.00789   0.00130  -0.00174
#>  3     3   0.99998  -0.00666  -0.00029  -0.00171
#>  4     4   0.99999  -0.00497  -0.00136  -0.00185
#>  5     5   0.99999  -0.00376  -0.00218  -0.00225
#>  6     6   0.99999  -0.00262  -0.00294  -0.00282
#>  7     7   0.99999  -0.00197  -0.00351  -0.00333
#>  8     8   0.99999  -0.00154  -0.00378  -0.00360
#>  9     9   0.99999  -0.00117  -0.00391  -0.00359
#> 10    10   0.99998  -0.00118  -0.00419  -0.00372
#> # ℹ 90 more rows
#> 
#> [[50]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99997  -0.00659   0.00207  -0.00181
#>  2     2   0.99998  -0.00587   0.00072  -0.00202
#>  3     3   0.99998  -0.00578  -0.00061  -0.00231
#>  4     4   0.99998  -0.00504  -0.00151  -0.00269
#>  5     5   0.99999  -0.00369  -0.00203  -0.00283
#>  6     6   0.99999  -0.00256  -0.00247  -0.00292
#>  7     7   0.99999  -0.00214  -0.00303  -0.00320
#>  8     8   0.99998  -0.00189  -0.00378  -0.00358
#>  9     9   0.99998  -0.00175  -0.00434  -0.00368
#> 10    10   0.99998  -0.00159  -0.00442  -0.00343
#> # ℹ 90 more rows
#> 
#> [[51]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99998  -0.00663   0.00021  -0.00243
#>  2     2   0.99996  -0.00807   0.00012  -0.00222
#>  3     3   0.99997  -0.00801  -0.00025  -0.00185
#>  4     4   0.99997  -0.00700  -0.00072  -0.00154
#>  5     5   0.99998  -0.00578  -0.00123  -0.00134
#>  6     6   0.99999  -0.00453  -0.00183  -0.00120
#>  7     7   0.99999  -0.00338  -0.00252  -0.00104
#>  8     8   0.99999  -0.00260  -0.00346  -0.00093
#>  9     9   0.99999  -0.00210  -0.00409  -0.00084
#> 10    10   0.99999  -0.00173  -0.00447  -0.00060
#> # ℹ 90 more rows
#> 
#> [[52]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99997  -0.00714   0.00038  -0.00254
#>  2     2   0.99997  -0.00783   0.00007  -0.00248
#>  3     3   0.99997  -0.00800  -0.00014  -0.00211
#>  4     4   0.99997  -0.00746  -0.00072  -0.00198
#>  5     5   0.99998  -0.00619  -0.00154  -0.00194
#>  6     6   0.99998  -0.00494  -0.00237  -0.00192
#>  7     7   0.99999  -0.00374  -0.00299  -0.00169
#>  8     8   0.99999  -0.00266  -0.00339  -0.00137
#>  9     9   0.99999  -0.00225  -0.00390  -0.00123
#> 10    10   0.99999  -0.00205  -0.00437  -0.00102
#> # ℹ 90 more rows
#> 
#> [[53]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99998  -0.00528   0.00375  -0.00126
#>  2     2   0.99998  -0.00616   0.00207  -0.00151
#>  3     3   0.99998  -0.00576   0.00039  -0.00163
#>  4     4   0.99999  -0.00477  -0.00101  -0.00189
#>  5     5   0.99999  -0.00371  -0.00193  -0.00212
#>  6     6   0.99999  -0.00306  -0.00271  -0.00244
#>  7     7   0.99999  -0.00240  -0.00345  -0.00283
#>  8     8   0.99998  -0.00189  -0.00408  -0.00319
#>  9     9   0.99998  -0.00169  -0.00435  -0.00334
#> 10    10   0.99998  -0.00159  -0.00442  -0.00335
#> # ℹ 90 more rows
#> 
#> [[54]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99998  -0.00512   0.00273  -0.00155
#>  2     2   0.99998  -0.00620   0.00183  -0.00167
#>  3     3   0.99998  -0.00547   0.00011  -0.00175
#>  4     4   0.99998  -0.00511  -0.00138  -0.00207
#>  5     5   0.99999  -0.00417  -0.00226  -0.00236
#>  6     6   0.99999  -0.00302  -0.00284  -0.00261
#>  7     7   0.99999  -0.00206  -0.00339  -0.00305
#>  8     8   0.99998  -0.00161  -0.00406  -0.00355
#>  9     9   0.99998  -0.00135  -0.00448  -0.00393
#> 10    10   0.99998  -0.00133  -0.00439  -0.00405
#> # ℹ 90 more rows
#> 
#> [[55]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99997  -0.00711   0.00274  -0.00228
#>  2     2   0.99997  -0.00772   0.00186  -0.00146
#>  3     3   0.99997  -0.00715   0.00136  -0.00106
#>  4     4   0.99998  -0.00623   0.00076  -0.00089
#>  5     5   0.99998  -0.00548  -0.00010  -0.00092
#>  6     6   0.99999  -0.00445  -0.00133  -0.00104
#>  7     7   0.99999  -0.00368  -0.00250  -0.00126
#>  8     8   0.99999  -0.00316  -0.00351  -0.00132
#>  9     9   0.99999  -0.00251  -0.00401  -0.00106
#> 10    10   0.99999  -0.00210  -0.00443  -0.00081
#> # ℹ 90 more rows
#> 
#> [[56]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99997  -0.00651   0.00286  -0.00281
#>  2     2   0.99997  -0.00763   0.00150  -0.00218
#>  3     3   0.99997  -0.00770   0.00107  -0.00162
#>  4     4   0.99997  -0.00707   0.00055  -0.00134
#>  5     5   0.99998  -0.00615  -0.00016  -0.00125
#>  6     6   0.99999  -0.00488  -0.00125  -0.00118
#>  7     7   0.99999  -0.00374  -0.00250  -0.00108
#>  8     8   0.99999  -0.00285  -0.00341  -0.00091
#>  9     9   0.99999  -0.00236  -0.00424  -0.00084
#> 10    10   0.99999  -0.00204  -0.00497  -0.00069
#> # ℹ 90 more rows
#> 
#> [[57]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99999  -0.00302   0.00113  -0.00151
#>  2     2   0.99999  -0.00336   0.00107  -0.00172
#>  3     3   0.99999  -0.00298   0.00091  -0.00154
#>  4     4   0.99999  -0.00325   0.00105  -0.00160
#>  5     5   0.99999  -0.00302   0.00089  -0.00172
#>  6     6   0.99999  -0.00284   0.00055  -0.00181
#>  7     7   0.99999  -0.00307   0.00055  -0.00188
#>  8     8   0.99999  -0.00277   0.00042  -0.00198
#>  9     9   0.99999  -0.00312   0.00048  -0.00194
#> 10    10   0.99999  -0.00309   0.00062  -0.00182
#> # ℹ 90 more rows
#> 
#> [[58]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99999  -0.00329   0.00247  -0.00075
#>  2     2   0.99999  -0.00338   0.00181  -0.00091
#>  3     3   0.99999  -0.00289   0.00129  -0.00112
#>  4     4   1.00000  -0.00218   0.00097  -0.00132
#>  5     5   1.00000  -0.00180   0.00062  -0.00161
#>  6     6   1.00000  -0.00167   0.00025  -0.00195
#>  7     7   1.00000  -0.00165  -0.00007  -0.00225
#>  8     8   1.00000  -0.00150  -0.00049  -0.00247
#>  9     9   0.99999  -0.00160  -0.00075  -0.00265
#> 10    10   0.99999  -0.00178  -0.00098  -0.00273
#> # ℹ 90 more rows
#> 
#> [[59]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99998  -0.00610   0.00042  -0.00038
#>  2     2   0.99999  -0.00525  -0.00010  -0.00038
#>  3     3   0.99999  -0.00385  -0.00110  -0.00046
#>  4     4   1.00000  -0.00246  -0.00169  -0.00073
#>  5     5   1.00000  -0.00142  -0.00232  -0.00117
#>  6     6   0.99999  -0.00092  -0.00323  -0.00151
#>  7     7   0.99999  -0.00071  -0.00401  -0.00183
#>  8     8   0.99999  -0.00053  -0.00453  -0.00205
#>  9     9   0.99999  -0.00019  -0.00470  -0.00202
#> 10    10   0.99999  -0.00024  -0.00475  -0.00198
#> # ℹ 90 more rows
#> 
#> [[60]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   1.00000  -0.00277  -0.00053  -0.00115
#>  2     2   1.00000  -0.00279  -0.00079  -0.00085
#>  3     3   1.00000  -0.00193  -0.00158  -0.00107
#>  4     4   1.00000  -0.00116  -0.00214  -0.00130
#>  5     5   0.99999  -0.00071  -0.00289  -0.00151
#>  6     6   0.99999  -0.00040  -0.00331  -0.00162
#>  7     7   0.99999  -0.00023  -0.00399  -0.00177
#>  8     8   0.99999  -0.00027  -0.00462  -0.00174
#>  9     9   0.99999  -0.00033  -0.00495  -0.00159
#> 10    10   0.99999  -0.00045  -0.00499  -0.00144
#> # ℹ 90 more rows
#> 
#> [[61]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99999  -0.00279   0.00336  -0.00123
#>  2     2   0.99999  -0.00318   0.00207  -0.00138
#>  3     3   0.99999  -0.00322   0.00125  -0.00173
#>  4     4   0.99999  -0.00300   0.00079  -0.00208
#>  5     5   0.99999  -0.00272   0.00064  -0.00240
#>  6     6   0.99999  -0.00233   0.00058  -0.00268
#>  7     7   0.99999  -0.00219   0.00048  -0.00302
#>  8     8   0.99999  -0.00223   0.00031  -0.00325
#>  9     9   0.99999  -0.00235   0.00020  -0.00351
#> 10    10   0.99999  -0.00263  -0.00002  -0.00379
#> # ℹ 90 more rows
#> 
#> [[62]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99999  -0.00416   0.00339  -0.00072
#>  2     2   0.99999  -0.00372   0.00279  -0.00121
#>  3     3   0.99999  -0.00296   0.00186  -0.00157
#>  4     4   0.99999  -0.00246   0.00123  -0.00201
#>  5     5   0.99999  -0.00226   0.00069  -0.00239
#>  6     6   0.99999  -0.00226   0.00039  -0.00266
#>  7     7   0.99999  -0.00236   0.00009  -0.00290
#>  8     8   0.99999  -0.00251  -0.00015  -0.00306
#>  9     9   0.99999  -0.00272  -0.00034  -0.00317
#> 10    10   0.99999  -0.00282  -0.00038  -0.00317
#> # ℹ 90 more rows
#> 
#> [[63]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99998  -0.00638   0.00002  -0.00025
#>  2     2   0.99999  -0.00494  -0.00044  -0.00074
#>  3     3   0.99999  -0.00409  -0.00155  -0.00097
#>  4     4   0.99999  -0.00321  -0.00229  -0.00124
#>  5     5   0.99999  -0.00244  -0.00278  -0.00154
#>  6     6   0.99999  -0.00193  -0.00330  -0.00184
#>  7     7   0.99999  -0.00160  -0.00362  -0.00196
#>  8     8   0.99999  -0.00147  -0.00398  -0.00203
#>  9     9   0.99999  -0.00150  -0.00431  -0.00206
#> 10    10   0.99999  -0.00165  -0.00459  -0.00212
#> # ℹ 90 more rows
#> 
#> [[64]]
#> # A tibble: 100 × 5
#>     time         w         x         y         z
#>    <int> <dec:.5!> <dec:.5!> <dec:.5!> <dec:.5!>
#>  1     1   0.99998  -0.00584   0.00114   0.00003
#>  2     2   0.99999  -0.00511  -0.00030  -0.00059
#>  3     3   0.99999  -0.00409  -0.00169  -0.00097
#>  4     4   0.99999  -0.00307  -0.00247  -0.00137
#>  5     5   0.99999  -0.00215  -0.00294  -0.00185
#>  6     6   0.99999  -0.00165  -0.00342  -0.00226
#>  7     7   0.99999  -0.00127  -0.00384  -0.00250
#>  8     8   0.99999  -0.00097  -0.00404  -0.00248
#>  9     9   0.99999  -0.00076  -0.00425  -0.00237
#> 10    10   0.99999  -0.00087  -0.00453  -0.00231
#> # ℹ 90 more rows
#> 
#> attr(,"class")
#> [1] "qts_sample" "list"