QTS Random Sampling

This function adds uncorrelated Gaussian noise to the logarithm QTS using an exponential covariance function.

Usage

rnorm_qts(n, mean_qts, alpha = 0.01, beta = 0.001)

Arguments

n

An integer specifying how many QTS should be generated.

mean_qts

An object of class qts specifying the mean QTS.

alpha

A positive scalar specifying the variance of each component of the log-QTS. Defaults to 0.01.

beta

A positive scalar specifying the exponential weight. Defaults to 0.001.

Value

A list of n objects of class qts with added noise as specified by parameters alpha and beta.

Details

See exp_cov_function for details about the roles of alpha and beta in the definition of the covariance operator.

Examples

rnorm_qts(1, vespa64$igp[[1]])
#> [[1]]
#> # A tibble: 101 × 5
#>     time     w     x         y      z
#>    <int> <dbl> <dbl>     <dbl>  <dbl>
#>  1     0 0.984 0.176  0.00976  0.0138
#>  2     1 0.986 0.165  0.00809  0.0213
#>  3     2 0.987 0.161  0.00994  0.0198
#>  4     3 0.988 0.155  0.00760  0.0231
#>  5     4 0.990 0.138  0.00668  0.0242
#>  6     5 0.991 0.128  0.00287  0.0249
#>  7     6 0.993 0.117  0.000760 0.0251
#>  8     7 0.994 0.109 -0.00137  0.0296
#>  9     8 0.993 0.115 -0.00349  0.0212
#> 10     9 0.993 0.120 -0.00696  0.0110
#> # … with 91 more rows
#> 
#> attr(,"class")
#> [1] "qts_sample" "list"